% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_family_dmt.R
\name{dmt}
\alias{dmt}
\title{Samples built upon cosine.}
\usage{
dmt(n, A, delta, delta_range = NULL, alpha, theta = NULL,
  wn = list(type = "rnorm", args = list(mean = 0, sd = 1), adjust = 0),
  seed = NULL)
}
\arguments{
\item{n}{The number of samples to create.}

\item{A}{A matrix whose first row should be the positions and
second row should be the probabilites of getting there.  The
cases of main interest to investigate is the case where this
describes a process having expectation \code{0}.}

\item{delta}{Either a single number, or a vector whose length
equals the number of columns in \code{A}.  This will be used as
the amplitudes for the cosine-terms.}

\item{delta_range}{Either \code{NULL} (the default) or a vector
similiar to \code{delta}.  When this is different from
\code{NULL} the effect will be that the values for the deltas
will be sampled uniformly based on the intervals that the
corresponding components of \code{delta} and \code{delta_range}
span.}

\item{alpha}{Either a single number, or a vector whose length equals
the number of columns in \code{A}.  This will be used as the
frequencies in the cosine-terms.}

\item{theta}{Either a single number, or a vector whose length
equals the number of columns in \code{A}.  This will be used as
the phase in the cosine-terms.  The default value \code{NULL}
will trigger a procedure that draws these numbers uniformly
between \code{0} and \code{2pi}.}

\item{wn}{A list that specifies how to sample from some
distribution.  The default is to use standard normal white
noise.  Note: It's assumed that this should be white noise, but
there's no test that prevents a user from selecting something
different.  The value \code{NULL} can be used to set this term
to zero.}

\item{seed}{An integer to be used for the specification of a seed,
the default value NULL will trigger a seed to be selected at
random.}
}
\value{
A sample of length \code{n} based on the specifications
    given in the other arguments
}
\description{
The \code{dmt}-function can create a sample where different
trigonometric functions live at separate levels.  This function is
at the moment rather simplistic, as its main intention is to just
create an example for ISNPS-2016 (and I'm running low on time)
}
\keyword{internal}
