% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rho.R
\name{find_rho}
\alias{find_rho}
\title{Estimate local correlation \eqn{$\rho$} between two variables.}
\usage{
find_rho(data, grid, b, .kernel = c("normal", "uniform"),
  return_arguments = FALSE)
}
\arguments{
\item{data}{A matrix containing the observations, one bivariate
observation in each row.}

\item{grid}{A matrix containing the points where we want to find
the local correlations, one bivariate point in each row.}

\item{b}{A vector containing the two bandwidhts to use.}

\item{.kernel}{Specification of the kernel to use, either
\code{normal} or \code{uniform}, with default being the
\code{normal} one if none is selected.  Note: The
\code{uniform}-kernel takes much longer to compute (a factor of
ten when used on a small sample of size \eqn{$n=200$}).}

\item{return_arguments}{Logical argument, default \code{FALSE},
that can be used to include the three arguments \code{data},
\code{grid} and \code{b} in the result from this function.}
}
\value{
The result of this function is a list.  For each grid-point
there will be an estimate of the local correlation and the
log-density of the corresponding local Gaussian approximation at
that point.  When \code{return_arguments} is \code{TRUE}, the three
arguments \code{data}, \code{grid} and \code{b} will also be
included.
}
\description{
This function use the local likelihood function proposed by Hjort &
Jones (1996) to find an estimate of the local correlation
\eqn{$\rho$} given \eqn{$n$} bivariate observations and a vector
\eqn{$b$} with the specified bandwidths.  Note that the code use
the simplifying assumption that the the marginals are standard
normal, which reduces the number of parameters to estimate from
five to one.  This function is based on code from H{\\aa}kon
Otneim.
}
\details{
In the computation, a product kernel that either is based
on the standard normal distribution or the uniform distribution is
used.  The default kernel is the normal one, since it seems to
behave better than the uniform kernel with regard to being able to
get a proper result (i.e. not \code{NA}) and since it also seems to
compute a bit faster when a computation is possible.  Note that the
code computes the required densities (related to the normal
distribution) directly instead of using \code{dnorm} and
\code{dmvnorm}, since that gives a speed gain.
}
\references{
Hjort, N. L., and Jones, M. C.: "Locally parametric
nonparametric density estimation.", The Annals of Statistics
(1996): 1619-1647.
}
