% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_approx_extract.R
\name{LG_approx_extract}
\alias{LG_approx_extract}
\title{Extract data needed for Local Gaussian Spectral Densities.}
\usage{
LG_approx_extract(approx_data, LG_type = c("par_one", "par_five"),
  .method = c("all", "mean", "median"))
}
\arguments{
\item{approx_data}{A list with Local Gaussian Approximations,
which in general contains more information than needed for the
computation of the Local Gaussian Spectral Densities.}

\item{LG_type}{One of \code{c("par_five", "par_one")}, i.e. should
the Local Gaussian Autocorrelations be based on the approach
using five parameters or the simplified approach using only one
parameter.  Note that even though \code{approx_data} contains
information about two cases, the code later on will only work
upon one case at the time (due to memory restrictions) and it's
thus necessary to specify the part of interest.}

\item{.method}{One or more of the values in \code{c("all", "mean",
"median")}, but the argument will only have an effect if the
length of the \code{content}-dimension in the \code{LG_type}
part of \code{approx_data} is larger than one.  The idea is
that this argument will be used with the values "mean" and
"median" while investigating blocks, and the value "all" when
doing a bootstrap-investigation.}
}
\value{
A list with two components, a list \code{info} and an array
    \code{pos_lag}. The list \code{info} contains stuff related to
    reproducibility, whereas the array \code{pos_lag} contains the
    required Local Gaussian Autocorrelations.
}
\description{
The estimated Local Gaussian Approximations, created by
\code{LG_approx} and \code{LG_boot_approx}, include a lot of
information that we don't need in the computation of the Local
Gaussian Spectral Densities.  In order to keep the code later on as
simple as possible, this function takes care of the extraction of
the relevant values, i.e. the Local Gaussian Correlations.
}

