% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_auto_cross_computations.R
\name{LG_auto_cross_computations}
\alias{LG_auto_cross_computations}
\title{Compute values based on the cross- and auto-spectra.}
\usage{
LG_auto_cross_computations(.list_array, .branch = c("on_diag",
  "off_diag"), .type = c("auto", "cross"), .spectrum = c("phase",
  "amplitude", "squared_coherence"))
}
\arguments{
\item{.list_array}{The list whose nodes contains the arrays of
interest.  Note, it is assumed that this list has depth two,
and that the following branches must be present: "auto/cross"
and "Re/Im" at respectively first an second level.}

\item{.branch}{Either "on_diag" or "off_diag".  This selects the
main branch of \code{.list_array} that is to be operated upon.
If unspecified, the default will be to use "on_diag".}

\item{.type}{Either "auto" or "cross", decides which main branch of
\code{.list_array} that is to be operated upon.  If
unspecified, the default will be to use "auto".}

\item{.spectrum}{One or more of the values "phase", "amplitude" and
"squared_coherence".}
}
\value{
The result of this function will be a list of arrays
    related to the combination of the arguments \code{.branch},
    \code{.type} and \code{.spectrum}.  The content of some of the
    lists are a bit misleading, as they are placeholders (without
    the correct content) that is needed in order for the code later
    on to work as required.  This is done in order to avoid wasting
    storage place on entities that can be computed directly later
    on. Moreover, this function is written with the intention of
    being used in a loop over the available arguments, so some
    combinations of the arguments (.i.e. "on_diag" and "auto") only
    contains placeholders witout any interesting content.
}
\description{
This function computes assorted spectra based on list-arrays
containing the cross- and auto-spectra, i.e. the phase-, amplitude
and squared coherence can be computed by this approach.  It is also
assumed that the arrays has been stored using the "folding" that
enables redundant information to be dropped, i.e. for the pairs
indexed "Yi_Yj" with "i" and "j" integers, it will always be
assumed that "i" is smaller than "j".  Finally, since this function
works as an internal helper in \code{LG_spectra}, it will also be
asumed that the list contains information about the variable pairs.
}
