% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_boot_spectra.R
\name{LG_boot_spectra}
\alias{LG_boot_spectra}
\title{Bootstrap-estimated statistics for Local Gaussian Spectral
Densities}
\usage{
LG_boot_spectra(LG_approx_extract_data, window, omega_vec, cut_vec,
  all_statistics = FALSE, log_ = FALSE, storage_only = FALSE, ...)
}
\arguments{
\item{LG_approx_extract_data}{The list created when
\code{LG_approx_extract} is used upon the Local Gaussian
Approximations, i.e. the restricted subset of data needed for
the computation of the local Gaussian spectra.  The list does
also specify what kind of Local Gaussian Spectral Densities
that should be created i.e. if it should be based on those from
one ore five parameters.}

\item{window}{The window-function(s) to use on the summands of the
local Gaussian spectra.  This argument will in general be
inherited from the one used upon the original time series.}

\item{omega_vec}{A vector of frequencies.  This arguments will in
general be inherited from the one used upon the original time
series, but note that it might be necessary to consider smaller
chunks in order to avoid out of memory problems.}

\item{cut_vec}{\code{vector}.  This argument can be inherited from
the one used on the original time series, but note that a
restriction to a subset might be a good idea -- in particular
if an inspection of the spectra (based on the original data)
indicates that the truncation point should be in that subset.}

\item{all_statistics}{A logical value, default \code{FALSE}.  The
default will return a bunch of quantiles based on the
collection of bootstrapped values, which can be used when we
need estimates for the confidence intervals.
\code{all_statistics} equals \code{TRUE} will in addition
return a bunch of statistics related to the distribution of the
bootstrapped values.}

\item{log_}{A logic argument with default value \code{FALSE}.  This
decides if the statistics should be computed based on the
logarithmic values of our replicates (when that makes sense,
i.e. when we have nonzero values).  It might be preferable to
use \code{log_=TRUE}, but as we have no guarantee that it
occasionally might not occur negative values for the estimated
local Gaussian spectra, the default has nevertheless been set
to \code{FALSE}.}

\item{storage_only}{A logic argument with default value
\code{FALSE}.  When this is used, only a \code{quote}-object
for a dummy-array of the correct format (dimension and
dimension-names) for the result are returned.  This will be
useful when we for some reason in advance of the actual
computation need to know the configuration of the
dimension-names.  Furthermore, when we see that the
computations of the local Gaussian spectra will require
intermediate objects too large for the memory to handle well --
but where the final array still will be of a tractable size
(since we only want to store a set of statistics based on the
bootstrapped local Gaussian spectra, not the whole bunch of
them) -- then the quote can be evaluated to a storage-array and
we can split the computation into smaller chunks whose results
are inserted into the final array one at a time.}

\item{...}{The \code{dotsMethods}, additional arguments that should
be delivered to \code{my_apply} can be specified here.  The
options being \code{.front}, \code{.parallel} and
\code{.cores}.  See the documentation of \code{my_apply} for
details.}
}
\value{
The result of this function will depend on the value of
    \code{storage_only}.  It will either be a quote for a
    placeholder array to be filled in when the computation are
    divided into smaller chunks, or it will be an array -- which
    always will contain a bunch of empirical quantiles (in order to
    get confidence intervals), and, if \code{all_statistics} equals
    \code{TRUE}, it will also contain the mean, standard deviation,
    skewness, kurtosis and mad (median absolute deviation).  These
    later values might perhaps not make too much sense when we
    investigate the spectra instead of the logarithms of them, but
    we have no guarantee that negative values will not occur and I
    thus think I will let it be for the time being.
}
\description{
Based on the local Gaussian spectra obtained from the
bootstrap-procedure, this function returns a variety of statistics.
The estimated confidence intervals are the primary motivation for
this function, but it can also return information about the
distribution of the bootstrap-estimated spectra.  This function
first wraps around \code{LG_spectra} in order to get the
bootstrap-estimated spectra, then it extract the statistics from
the resulting array.  This function is a part of the bootstrap
wrapper, that attempts to ensure that the arguments delivered to it
are in small enough chunks to avoid out of memory errors.
}

