% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_boot_spectra_scribe.R
\name{LG_boot_spectra_scribe}
\alias{LG_boot_spectra_scribe}
\title{Bootstrap-estimated statistics for local Gaussian spectral
densities, scribe-function}
\usage{
LG_boot_spectra_scribe(main_dir, data_dir, spectra_dir, LG_type = NULL,
  omega_vec = NULL, window = NULL, cut_vec = NULL,
  all_statistics = FALSE, log_ = FALSE, threshold = 500)
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{data_dir}{A specification of the directory to be used when
loading and saving data.}

\item{spectra_dir}{A directory containing Local Gaussian Spectral
Densities based upon the original data from \code{data_dir}.
This is included in order to sanity-check the arguments and in
order to ensure that the info-file keep track of the
connections in the file-hierarchy.}

\item{LG_type}{One of \code{c("par_five", "par_one")}, i.e. should
the Local Gaussian Autocorrelations be based on the approach
using five parameters or the simplified approach using only one
parameter.  When \code{NULL} is used, the function will use the
\code{LG_type}-argument from the original local Gaussian
spectral analysis.}

\item{omega_vec}{A vector of frequencies.  This must be a subset of
the values used when analysing the original time series.  If
the default value \code{NULL} is used, then the function will
look up the \code{omega_vec}-argument from the original local
Gaussian spectral analysis.}

\item{window}{The window-function(s) to use on the summands of the
local Gaussian spectra. This must be a subset of the values
used when analysing the original time series.  If the default
value \code{NULL} is used, then the function will look up the
\code{window}-argument from the original local
Gaussian spectral analysis.}

\item{cut_vec}{The truncation-points to be used for the functions
selected in \code{window}.  This must be a subset of the values
used when analysing the original time series.  If the default
value \code{NULL} is used, then the function will look up the
\code{cut_vec}-argument from the original local Gaussian
spectral analysis.  A restriction of \code{cut_vec} to a
relevant subset (based on an inspection of the original result)
might be a good idea, since the computational time then can be
reduced to a fraction.}

\item{all_statistics}{A logical value, default \code{FALSE}.  The
default will return a bunch of quantiles based on the
collection of bootstrapped values, which can be used when we
need estimates for the confidence intervals.
\code{all_statistics} equals \code{TRUE} will in addition
return a bunch of statistics related to the distribution of the
bootstrapped values.}

\item{log_}{A logic argument with default value \code{FALSE}.  This
decides if the statistics should be computed based on the
logarithmic values of our replicates (when that makes sense,
i.e. when we have nonzero values).  It might be preferable to
use \code{log_=TRUE}, but as we have no guarantee that it
occasionally might not occur negative values for the estimated
local Gaussian spectra, the default has nevertheless been set
to \code{FALSE}.}

\item{threshold}{An integer, default value 500 (measured in MB),
that tells the program when a computation should be divided
into smaller chunks.  This reduces the chance of memory-related
problems, but as the present incarnation of \code{LG_splitter}
are rather stupid such problems might still occur for long
time-series where a huge number of lags are included.}
}
\value{
This function will return a two component list to the
    work-flow.  The first component is the logic value
    \code{done_before} that reveals whether or not the result
    already existed (in the specified file-structure), whereas the
    second component \code{data_dir} gives the location of the
    saved data.
}
\description{
This function takes care of the bookkeeping when we use
bootstrap-techniques to investigate the local Gaussian spectra of a
time series.
}
\details{
This function records its arguments and compares them to a
    previously stored information-object, in order to avoid redoing
    previously performed computations.  For new computations,
    relevant data will be extracted from \code{data_dir}, which
    then will be analysed in order to see if memory issues requires
    that the computations should be performed in smaller chunks.
    The function then calls \code{LG_approx_extract} and
    \code{LG_boot_spectra} in order to obtain the result -- and
    finally the result is added to the file-structure while the
    information-object is updated.
}

