% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_collect_blocks.R
\name{LG_collect_blocks}
\alias{LG_collect_blocks}
\title{Collect data from block of independent series into one object.}
\usage{
LG_collect_blocks(main_dir, data_dir, threshold = 500)
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{data_dir}{The directory containing the confidence intervals
part from the function \code{LG_Wrapper_Blocks}.}

\item{threshold}{An \code{integer}, default value 500 (measured in
MB), that gives the upper in-memory size of the objects stored
in the resulting files.}
}
\value{
A path to be used by \code{LG_shiny}.
}
\description{
The computation of the local Gaussian spectra for the blocks of
independent series creates a file-hierarchy that differs a bit from
the setup of the resulting code.  In order to allow the present
incarnation of \code{LG_shiny} to get access to the results, some
tweaking of the related objects must be done.
}
\details{
Based on the \code{data_dir} argument, there will be
    performed an adjustment of the corresponding file-hierarchy --
    in order for \code{LG_shiny} to find the data (without a major
    rewriting of the code in that function.)
}
