% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_create_plot_df.R
\name{LG_create_plot_df}
\alias{LG_create_plot_df}
\title{Create the data needed when plotting the local Gaussian
correlations and spectra.}
\usage{
LG_create_plot_df(.look_up, ..env)
}
\arguments{
\item{.look_up}{The list created by \code{LG_lookup}, containing
the information needed in order to decide what kind of
data-frame that is required.}

\item{..env}{The environment containing the required information,
and also the environment where the resulting data-frame will be
stored.}
}
\value{
The purpose of this function is to create a list with the
    data needed for the function \code{LG_plot}.  The result will
    be stored in the environment \code{..env}, and the unique name
    (computed from the input parameters used in the construction of
    it) will then be returned to the workflow.  This solution
    enables a simple form of caching, since the computations only
    will be performed when required.
}
\description{
This function created a list with the data needed for the plots of
the estimated local Gaussian correlations and spectra.  A simple
caching is performed in order to avoid having computations more
than once.
}
\keyword{internal}
