% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_default.R
\docType{data}
\name{LG_default}
\alias{LG_default}
\title{Internal default values for the local Gaussian machinery}
\format{list}
\usage{
This list is used as an internal reference for the
    functions in this package.
}
\description{
This list contains internal default-values that other
    functions in the package need in order to "communicate"
    together.  This list governs the naming conventions used for
    the dimension-names for arrays; it specifies the
    directory-hierarchy and the prefixes to be used for the files
    stored in them; and it also specifies assorted information that
    the sanity-check function need to have access to.
}
\keyword{internal}
