% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_explain_plot.R
\name{LG_explain_plot}
\alias{LG_explain_plot}
\title{Explain the content of a plot.}
\usage{
LG_explain_plot(.plot_details, .mode = NULL, .digits_for_points = 2,
  .digits_for_percentiles = 0)
}
\arguments{
\item{.plot_details}{This can either be a saved plot created by this
package, or it can be the details-list directly.  (The latter
alternative is included to simplify the use of this function
inside of the internal \code{shiny}-application.)}

\item{.mode}{Specify the mode that the result will be returned in.
Two modes are supported, "markdown" and "latex".  The default
value \code{NULL} will trigger the selection of "latex" if
\code{.plot_details} is given as a saved \code{ggplot2}-object,
whereas "markdown" will be selected when \code{.plot_details}
is given as a list.}

\item{.digits_for_points}{An integer that specifies the number of
decimals to include when presenting a point.  The default value
is \code{2}.}

\item{.digits_for_percentiles}{An integer that specifies the number
of decimals to include when presenting the percentiles (of the
standard normal distribution) that corresponds to the point.
The default value is \code{0}.}
}
\value{
The returned result will be a description of the plot in
    \code{.plot_details}, in the mode specified by \code{.mode}.
}
\description{
This function gives information about the different plots created
by this package.  It is used internally in the
\code{shiny}-application, and it can also be used when a plot is
desired included in an article/presentation.  In the latter case
the plot must first be saved to an object, and that object can then
be given to this function in the argument \code{.plot_details}.
Note that the returned text is based on the information that the
saved plot-object has in the attribute named \code{details}, and it
is thus of course possible to create customised presentations
directly based on the information found there.
}
