% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_extend_points.R
\name{LG_extend_points}
\alias{LG_extend_points}
\title{Extend select points for the local Gaussian inspections}
\usage{
LG_extend_points(LG_points)
}
\arguments{
\item{LG_points}{The result of the function \code{LG_select_points}.}
}
\value{
An extended version of the array from \code{LG_points} will
    be returned, with column names "v1" and "v2" and row names
    which identifies the points of interest (to be used in the
    argument grids).  The returned array has been screened for
    duplicates, and only contains unique points.
}
\description{
This internal function adds diagonally related points
    to the computation of the local Gaussian (auto- and cross-)
    correlations.  If a point (v1,v2) is present, then this
    function will ensure that the three points (v1,v1), (v2,v2) and
    (v2,v1) also will be included in the analysis.
}
\keyword{internal}
