% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_extract_df.R
\name{LG_extract_df}
\alias{LG_extract_df}
\title{Final computations of assorted spectra.}
\usage{
LG_extract_df(.list_array, .global = TRUE, .diag = c("on_diag",
  "off_diag"), .spectrum = c("auto", "co", "quad", "phase", "amplitude",
  "squared_coherence"), .adjust_result = FALSE, .branch = list(),
  .Y_range = NULL)
}
\arguments{
\item{.list_array}{The list whose nodes contains the arrays of
interest.}

\item{.global}{Logical value, default \code{TRUE}, which decides if
the operation is the one for the ordinary (global) spectral
densities, or if a variant needed for the local Gaussian
versions should be employed.  (This argument is needed since
the depth of the \code{.list_array} is different for the local
and global data.)}

\item{.diag}{One of the values "on_diag", "off_diag".  Only used
when \code{.global} is \code{FALSE}, i.e. it will for the local
Gaussian cases specify which branch(es) of \code{.list_array}
that should be used in the computation.}

\item{.spectrum}{One of the values "auto", "co", "quad", "phase",
"amplitude", "squared_coherence".  This specifies the
extraction procedure to be performed.}

\item{.adjust_result}{Logical value, default \code{FALSE}, which
can be used when there is a need for some minor tweaking of the
result to get the correct version.  This is needed since
"folding" has been used on the saved data to reduce the
memory-load, so for some extractions it will be necessary to
tweak the result a bit.  When this argument is \code{TRUE}, the
effect will be a change of sign when \code{.spectrum} is equal
to "quad" and "phase".}

\item{.branch}{A list of logical values needed for those cases
where off-diagonal points are encountered.}

\item{.Y_range}{A vector containing to integers, which can be used
when it is desirable to compute a restricted range where the
most extreme observations have been ignored.  The default value
\code{NULL} takes the full range.}
}
\value{
The result of this function will either be a restricted
    range (when \code{.Y_range} has been specified, or else it will
    be a list with two nodes, one node being the full range and the
    other node a data-frame (to be used for the plot-functions).
}
\description{
This function takes care of the extraction (and final adjustment)
needed in \code{LG_plot_load} when the list-arrays loaded from
different files is to be modified to a data frame which can be
digested by the graphical functions.
}
\keyword{internal}
