% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_extract_scripts.R
\name{LG_extract_scripts}
\alias{LG_extract_scripts}
\title{Extraction of scripts.}
\usage{
LG_extract_scripts(target_dir)
}
\arguments{
\item{target_dir}{The path to the directory where the scripts are
to be added.  Note that the target directory must exist and be
empty in order for this function to work.  The
existence-requirement is included in order to avoid that
directories are accidentally created, and the empty-requirement
is included in order to avoid that previously extracted (and
potentially modified) versions of the scripts are overwritten.}
}
\value{
If successful, the scripts are moved into \code{target_dir},
    and the user can then run them (or modify them) as desired.
}
\description{
This function extracts the scripts from the internal file hierarchy
to a target directory selected by the user.
}
