% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_load.R
\name{LG_load}
\alias{LG_load}
\title{Load an object into an environment}
\usage{
LG_load(.file, .env = parent.frame(), .name = NULL, .attributes_extra = NULL)
}
\arguments{
\item{.file}{The path to the file to be loaded.  An error will be
returned if more than one object is contained in \code{.file}.}

\item{.env}{The environment, default \code{parent.frame()}, where
the result should be added under the name given to the argument
\code{.name}.}

\item{.name}{A character-string giving the name to be used for the
object loaded from \code{.file}.  The default value \code{NULL}
will trigger the original name of the loaded object, i.e. the
name it was saved under in \code{.file}.}

\item{.attributes_extra}{A list containing additional attributes to
be added to the loaded object.  The default value \code{NULL}
will skip this part.}
}
\value{
An object named \code{.name} will be added from
    \code{.file} to \code{.env}, with the possible extension of
    attributes as given by \code{.attributes_extra}.
}
\description{
This internal function loads the content of a file
    into an environment, and it gives it the name specified in the
    \code{.name}-argument.  It is required that the file only
    contains one single object.
}
\keyword{internal}
