% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup.R
\name{LG_lookup}
\alias{LG_lookup}
\title{Extract details from the logged information.}
\usage{
LG_lookup(input, .AB_env)
}
\arguments{
\item{input}{A non-reactive version of the \code{input}-list
generated by \code{NEW_LG_shiny}.}

\item{.AB_env}{The environment that contains the desired
information.}
}
\value{
This function will extract information from the logged
    information, in a format suited to the needs of the the other
    functions involved in the interactive investigation of the
    different plots that can be created from the estimated local
    Gaussian correlations.  The resulting list will be based on the
    \code{input}-argument, and it will at the top-level contain
    some derived values (e.g. logical values) that will be used in
    other functions to avoid redoing code many places.  The list
    will in addition contain a node named \code{details} that will
    be added as an attribute to the resulting plot, and this node
    will contain text that describes the content of the plot.  The
    result will also contain a \code{cache}-node, that contains the
    names needed for the caching of some intermediate results in
    the computations.
}
\description{
This function creates a list that contains different values from
the logged structure.  The purpose of this is to collect most of
the pesky details in one function, so other functions can avoid
these details.
}
\keyword{internal}
