% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_cache.R
\name{LG_lookup_cache}
\alias{LG_lookup_cache}
\title{Helper-function to create the \code{cache}-list of \code{look_up}}
\usage{
LG_lookup_cache(look_up)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}
}
\value{
A \code{cache}-list will be returned to the calling
    function (i.e. \code{LG_lookup}), and this list contains the
    cache-keys that are used in order to avoid that already
    performed tasks are recomputed.
}
\description{
This internal helper function prepares the unique
    cache-keys that enables a primitive caching-procedure to be
    used in the interactive \code{LG_shiny}-application.
}
\keyword{internal}
