% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_curlicues.R
\name{LG_lookup_curlicues}
\alias{LG_lookup_curlicues}
\title{Manage the curlicues to be used on the plots}
\usage{
LG_lookup_curlicues(look_up)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}
}
\value{
A list with the curlicues-information to be used in the
    plot, e.g. details related to the size of lines, and the
    position, colour and size of the annotated text
}
\description{
Manage the curlicues to be used on the plots
}
\keyword{internal}
