% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_curlicues_plot_stamp.R
\name{LG_lookup_curlicues_plot_stamp}
\alias{LG_lookup_curlicues_plot_stamp}
\title{Create the plot-stamp-label for the the curlicues list}
\usage{
LG_lookup_curlicues_plot_stamp(look_up, CSC)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}

\item{CSC}{The case-specific-curlicues (CSC) from the internal
workings of the \code{LG_lookup_curlicues}-function.}
}
\value{
This function will create the plot-stamp-label that should
    be added to the \code{CSC}-list in the calling function.
}
\description{
Create the plot-stamp-label for the the curlicues list
}
\keyword{internal}
