% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_details.R
\name{LG_lookup_details}
\alias{LG_lookup_details}
\title{Helper-function to create the \code{details}-list of \code{look_up}}
\usage{
LG_lookup_details(look_up, .AB_env)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}

\item{.AB_env}{The environment that contains the desired
information.  This argument is inherited from \code{LG_lookup}.}
}
\value{
A \code{details}-list will be returned to the calling
    function (i.e. \code{LG_lookup}), and this list contains the
    details needed for the description of the content of the plots.
}
\description{
This internal helper function prepares the information
    that is needed in order to describe the content of a given
    plot.  This information will be added as an attribute when a
    plot is created in a nonreactive setting, which implies that it
    is easily available when a plot is to be included in a
    paper/presentation.  See \code{LG_explain_plots} for further
    details.
}
\keyword{internal}
