% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_details_NC_fail.R
\name{LG_lookup_details_NC_fail}
\alias{LG_lookup_details_NC_fail}
\title{Helper-function to deal with the details when the numerical convergence fails.}
\usage{
LG_lookup_details_NC_fail(look_up, .AB_env)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}

\item{.AB_env}{The environment that contains the desired
information.  This argument is inherited from \code{LG_lookup}.}
}
\value{
An additional node will be added to the \code{details}-list
    in the calling function \code{LG_lookup_details}.  This list
    can be used to figure out if the observed issue regarding
    failed numerical convergence is present for the case under
    investigation.
}
\description{
This internal helper function is only called when one
    (or more) of the estimated local Gaussian (auto- or cross-)
    correlations fails to converge numerically.  In this case the
    underlying convergence information will be analysed in order to
    identify more details about the problem.
}
\keyword{internal}
