% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_restrict.R
\name{LG_lookup_restrict}
\alias{LG_lookup_restrict}
\title{Helper-function to create the restrict-list}
\usage{
LG_lookup_restrict(look_up = look_up)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}
}
\value{
A \code{restrict} list will be returned to the calling
    function (i.e. \code{LG_lookup}), and this list contains the
    details needed for the slicing and dicing of the data loaded
    from file.
}
\description{
Helper-function to create the restrict-list
}
\keyword{internal}
