% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_merge_files.R
\name{LG_merge_files}
\alias{LG_merge_files}
\title{Merge files containing data created by different steps.}
\usage{
LG_merge_files(components, source_dir, save_dir)
}
\arguments{
\item{components}{A list containing information that identifies the
names of the files (in the directories specified by
\code{source_dir}) that contains the data to be loaded, and
also names to be used on the files where the merged that will
be saved (in the directory specified by \code{save_dir})}

\item{source_dir}{A list with information related to the paths
needed to find the directories where the data of interest are
stored.}

\item{save_dir}{The information needed to find the directory where
the merged files will be saved (with names as given in
\code{components}).}
}
\value{
The primary goal of this function is to merge the files,
    and save them at the assigned positions in the file-hierarchy.
    When this is done, the \code{components} object will be updated
    with information about the sizes (in MB) of the saved objects,
    some excess information will be pruned away, and the revised
    object will then be returned to the workflow (in order for the
    calling function to update the \code{info}-object with it).
}
\description{
Merge files containing data created by different steps.
}
\details{
Some plots depends on content from different parts of the
    overall algorithm, for these cases it's of interest to merge
    the relevant data into new files.  This process is not straight
    forward, since the data from the different steps might be
    distributed over several files.  (This function will most likely
    be an internal function in the package, i.e. that it will not
    be exported.)
}
