% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_normalisation_adjustment.R
\name{LG_normalisation_adjustment}
\alias{LG_normalisation_adjustment}
\title{Create normalised time series from the original ones.}
\usage{
LG_normalisation_adjustment(TS, .normalisation_rule = c("ecdf", "logspline"),
  .adjustment_rule = "data", .remove_ties = TRUE)
}
\arguments{
\item{TS}{The time series object to normalise.  It's assumed that
this is structured in an array, with one dimension-name being
"content".}

\item{.normalisation_rule}{A value from \code{c("ecdf",
"logspline")} that decides which kind of normalisation rule to
use upon the data.  The "logspline"-alternative seems to give a
more faithful representation of the data, but can encounter
problems if there's "to much data" in some regions (which can
occur for a long time series if some really extreme
observations are present).  The "ecdf" doesn't encounter such
problems, and it will also be much quicker to compute, which
for larger time series might be an important aspect to keep in
mind.}

\item{.adjustment_rule}{Either a non-negative number, or "data".
This will be added as an attribute to the result, and later on
it will decide if any finite-sample adjustment should be used
for the estimated local Gaussian autocorrelations.  Note that
no adjustments will be performed when
\code{.adjustment_rule=0}.}

\item{.remove_ties}{A logical value, default \code{TRUE}, in which
case the presence of ties will trigger a minor perturbation of
the data.  Note: Whenever this happens, \code{set.seed(1)}
will be used in order to ensure reproducibility.}
}
\value{
A normalised version of the time series, with attributes
    specifying how to adjust later on due to the effect of finite-
    samples.
}
\description{
Create normalised time series from the original ones.
}
\details{
A normalisation of the time series under consideration is
    preferable when an analysis based on the local Gaussian
    approximations are of interest.  The intention of this function
    is to allow for easier adjustment of the normalisation regime,
    i.e if the normalisation should be based on the cumulative
    density function, the logspline-density (or other options that
    might be added later on).  This function will also add some
    attributes to the result, with the required instruction for the
    finite-sample adjustments to be used when computing the local
    Gaussian autocorrelations.
}

