% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot.R
\name{LG_plot}
\alias{LG_plot}
\title{Create plots of local Gaussian correlations and spectral densities}
\usage{
LG_plot(..env, look_up)
}
\arguments{
\item{..env}{The environment with the information needed in order
to create the desired plot.}

\item{look_up}{The environment created by \code{LG_lookup}, which
keep tracks of information for a wide range of
helper-functions.}
}
\value{
A plot is returned to the work-flow.
}
\description{
This internal function creates the plots of the local
    Gaussian correlations and spectral densities.
}
\note{
The function \code{LG_plot_helper} is the one that should be
    used when it is of interest to reproduce the plots in a
    non-reactive environment, in particular since it is there it is
    possible to fine-tune the size of the annotated text.  This
    latter detail is of importance when several plots are to be
    included in a grid, see the scripts for examples (use
    \code{LG_extract_scripts} in order to access them).
}
\keyword{internal}
