% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_complex.R
\name{LG_plot_complex}
\alias{LG_plot_complex}
\title{Complex-valued-plots for the local Gaussian spectra}
\usage{
LG_plot_complex(..env, look_up, .selection = c("polar", "Cartesian",
  "zoom"))
}
\arguments{
\item{..env}{The environment containing the desired information
from which the data should be extracted}

\item{look_up}{The environment containing the details needed for
the investigation.}

\item{.selection}{Specifies the type of plot to be used.  The value
should be one of \code{c("polar", "Cartesian", "zoom")}.  (This
should be a temporary argument, since it later on should be
defined as a part of the shiny-interface, and thus occur in the
\code{look_up}-environment.)}
}
\value{
A plot is returned to the workflow.
}
\description{
This internal function plots the complex-valued
    spectra (for a given frequency) in the complex plane.
}
\keyword{internal}
