% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_curlicues.R
\name{LG_plot_curlicues}
\alias{LG_plot_curlicues}
\title{Keep control over the curlicues to be used on the plots}
\usage{
LG_plot_curlicues(new_curlicues = NULL, old_curlicues = NULL)
}
\arguments{
\item{new_curlicues}{A list that contains values that should be
used to update the content of \code{old_curlicues}.}

\item{old_curlicues}{The list to be updated with the content from the list
given in \code{new_curlicues}.}
}
\value{
The result of this function will be that the information in
    \code{old_curlicues} will be updated based on the information
    from \code{new_curlicues}.  If both arguments are equal to the
    default value \code{NULL}, then te result will be a copy of the
    default values stored in \code{LG_default}.
}
\description{
Keep control over the curlicues to be used on the plots
}
\keyword{internal}
