% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_df_spectra.R
\name{LG_plot_df_spectra}
\alias{LG_plot_df_spectra}
\title{Extract data-frames needed for the plotting of the spectra}
\usage{
LG_plot_df_spectra(look_up, ..env)
}
\arguments{
\item{look_up}{An environment created by \code{LG_lookup} (in the
function \code{LG_plot_helper}), where the key details have
been extracted (or constructed) from a (non-reactive copy) of
the values defined by the \code{shiny}-interface.}

\item{..env}{The environment that contains the correlations, and
also the environment in which the result will be stored.}
}
\value{
An update of \code{..env} will be performed with values
    needed for the plotting of the spectra.
}
\description{
This internal function computes the spectra for a
    collection of (local Gaussian) auto- and cross-correlations.
    This function is an internal function which is called from the
    functions that creates the plots.
}
\keyword{internal}
