% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_heatmap.R
\name{LG_plot_heatmap}
\alias{LG_plot_heatmap}
\title{Heatmap-plots for local Gaussian correlations and spectra}
\usage{
LG_plot_heatmap(..env, look_up)
}
\arguments{
\item{..env}{The environment containing the desired information
from which the data should be extracted}

\item{look_up}{The environment containing the details needed for
the investigation.}
}
\value{
A heatmap-plot is returned to the workflow.
}
\description{
This internal function creates heatmap-based plots
    based on the estimated local Gaussian correlations and spectra.
}
\keyword{internal}
