% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_helper.R
\name{LG_plot_helper}
\alias{LG_plot_helper}
\title{Helper-function for \code{LG_plot}.}
\usage{
LG_plot_helper(main_dir, input, input_curlicues = NULL, .env)
}
\arguments{
\item{main_dir}{The part of the \code{data_dir}-argument from the
function \code{LG_shiny}, that specifies the main path to where
the information is stored.  Note that this argument only is
required when the function is called outside of the shiny
application.  (Inside of the shiny-application, the relevant
information will be extracted from \code{.env}.)}

\item{input}{A list with the arguments needed in order to extract
the desired data for the plot of interest.}

\item{input_curlicues}{A list that can be used when this function
is used outside of the \code{shiny}-application.  This can be
used to fine-tune details related to the annotated information.
It is also possible to drop annotation from the plot, which
might be of interest when many plots are to be included in a
grid-based setup.  The default value \code{NULL} will ensure
that the plot will be created with the same setting that was
used in the \code{shiny}-application (this might however not
produce decent-looking plots when a grid of plots is created
and saved to file).}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}
}
\value{
The default behaviour is that a plot-object will be
    returned, with attributes describing the content. The
    information can be extracted from the plot-object by the help
    of \code{LG_explain_plot}.
}
\description{
This function extracts the desired array of local
    Gaussian (auto- or cross-) correlations from file, and then it
    sends those to the relevant plot-function in order to create
    the plot of interest.
}
\details{
This function is used as a helper in in the interactive
    setup of \code{LG_shiny}, but it can also be called in a
    non-reactive environment.  The interactive interface can be
    used to trigger the creation of the code needed in order to use
    this function in a non-reactive setting (which can be of
    interest e.g. when an example should be used in a
    paper/presentation).

The plots created by this function when it is used in a
    non-reactive setting, will contain an attribute \code{details}
    that can be used to extract information about the content.  The
    function \code{LG_extract_details} can be used to obtain the
    same information that was available in the interactive
    interface, or the user can extract the components of interest
    in order to produce other presentations too.
}
