% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_helper_extract_data_only.R
\name{LG_plot_helper_extract_data_only}
\alias{LG_plot_helper_extract_data_only}
\title{Extract only data for the plots}
\usage{
LG_plot_helper_extract_data_only(main_dir, input, input_curlicues = NULL,
  .env)
}
\arguments{
\item{main_dir}{The part of the \code{data_dir}-argument from the
function \code{LG_shiny}, that specifies the main path to where
the information is stored.  Note that this argument only is
required when the function is called outside of the shiny
application.  (Inside of the shiny-application, the relevant
information will be extracted from \code{.env}.)}

\item{input}{A list with the arguments needed in order to extract
the desired data for the plot of interest.}

\item{input_curlicues}{A list that can be used when this function
is used outside of the \code{shiny}-application.  This can be
used to fine-tune details related to the annotated information.
It is also possible to drop annotation from the plot, which
might be of interest when many plots are to be included in a
grid-based setup.  The default value \code{NULL} will ensure
that the plot will be created with the same setting that was
used in the \code{shiny}-application (this might however not
produce decent-looking plots when a grid of plots is created
and saved to file).}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}
}
\value{
This internal function will extract the underlying data
    from the plot created by the \code{LG_plot_helper}-function.
    The result is given as a list containing the following two
    nodes.

\describe{

\item{..env}{An environment containing the underlying data.}

\item{look_up}{A list containing the chunks of information that the
    internal functions need in order to extract the correct data
    for the specified plot.  This list does also include details
    needed in order for the explanation of the content of the
    plots.}

}
}
\description{
This internal function is used to extract the
    underlying data for a plot created by \code{LG_plot_helper}.
}
\details{
This function is primarily a helper function used during
    the development of new plots.  Some of the scripts might also
    use this function, but that is intended as a temporary
    solution.
}
\keyword{internal}
