% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_load.R
\name{LG_plot_load}
\alias{LG_plot_load}
\title{Load files to be used during the interactive inspection.}
\usage{
LG_plot_load(look_up, .env, .extract_LG_data_only)
}
\arguments{
\item{look_up}{A list created by \code{LG_lookup} (in the function
\code{LG_plot_helper}), where the key details have been
extracted (or constructed) from a (non-reactive copy) of the
values defined by the \code{shiny}-interface.}

\item{.env}{The environment in which the loaded stuff should be
stored.}

\item{.extract_LG_data_only}{Logical argument, inherited from
\code{LG_plot_helper}.  The value \code{FALSE} will ensure that
a plot is created, whereas the value \code{TRUE} is used when
the desired outcome is an object containing the estimated local
Gaussian values.}
}
\value{
The required data will be loaded from files into
    \code{.env}, but only if it hasn't been done before.  The
    resulting array will then be assigned to the environment of the
    calling function under the name given by \code{.result}.
}
\description{
This helper function that takes care of some details related to the
loading of files, i.e. with regard to the interactive inspection
required by \code{shiny}.
}
\keyword{internal}
