% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_TS.R
\name{LG_plot_TS}
\alias{LG_plot_TS}
\title{Plot a (subset of a) time series.}
\usage{
LG_plot_TS(.env)
}
\arguments{
\item{.env}{The environment containing the required details for the
extraction and presentation of the sample of interest.}
}
\value{
A plot of the time series under investigation will be
    returned.
}
\description{
This helper function creates a simple plot of the time
    series under investigation, and some minor details related to
    the points of investigation can in addition also be added.
}
\keyword{internal}
