% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_sanity_cheks.R
\name{LG_sanity_checks}
\alias{LG_sanity_checks}
\title{Sanity-check the arguments of the major exported functions}
\usage{
LG_sanity_checks(spy_report)
}
\arguments{
\item{spy_report}{The result returned by \code{spy} when used on
the targeted function.}
}
\value{
The execution will be stopped if any erroneous arguments
    are detected, in which case information about the problem(s)
    will be returned to the workspace.  If no problems are found,
    the programs returns nothing.  Note that this function works
    upon a shortcut to the environment part of \code{spy_report},
    and this implies that any adjustments performed by this
    function are registered in the corresponding
    \code{spy_report}-environment in the calling function.
}
\description{
This internal function checks that the arguments of
    some of the major exported functions are valid, and it will
    also adjust/standardise some of the values in the environment
    of the function under investigation.
}
\keyword{internal}
