% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny.R
\name{LG_shiny}
\alias{LG_shiny}
\title{Shiny app for a local Gaussian inspection of time series.}
\usage{
LG_shiny(main_dir, data_dir = NULL, default_type = "par_five")
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{data_dir}{The directory containing the data of interest.
When this argument is \code{NULL}, then the investigation
starts at the top level without attempting to select a
particular part of the file-hierarchy.}

\item{default_type}{Specifies whether "par_five" or "par_one" will
be used, i.e. if the local Gaussian auto-correlation should be
based on an approximation based on five or one parameters.
This argument will be ignored when only one alternative is
present.}
}
\value{
An interactive shiny application will be loaded (if
    \code{data_dir} points to a relevant directory containing the
    desired data).
}
\description{
This function opens an interactive Shiny-application that can be
used for an inspection of the stored data.  If I succeed in what I
want to do, the function should also be able to update the
corresponding information, such that functions later on can take as
arguments stuff saved from a previous interactive inspection.
}
