% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_C_input.R
\name{LG_shiny_C_input}
\alias{LG_shiny_C_input}
\title{Create the correlation-specific part of the \code{LG_shiny}-interface.}
\usage{
LG_shiny_C_input(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function removes those parts of the
    \code{LG_shiny}-interface that is superfluous when the target
    of the investigation is the correlations.
}
\description{
This internal function is used to update the interface
    created by \code{LG_shiny}, in order for the investigation to
    focus on the local Gaussian (auto- and cross-) correlations.
}
\keyword{internal}
