% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_correlation.R
\name{LG_shiny_correlation}
\alias{LG_shiny_correlation}
\title{Extraction of the correlations.}
\usage{
LG_shiny_correlation(look_up, ..env)
}
\arguments{
\item{look_up}{A list created by \code{LG_lookup} (in the function
\code{LG_plot_helper}), where the key details have been
extracted (or constructed) from a (non-reactive copy) of the
values defined by the \code{shiny}-interface.}

\item{..env}{The environment that contains the correlations, and
also the environment in which the result will be stored.}
}
\value{
The environment \code{..env} will be updated with values
    needed for the plotting of the spectra.  Internal caching is
    used in order to avoid repetition of already performed tasks.
    The creation of data-frames are taken care of in
    \code{LG_create_plot_df}.
}
\description{
This helper function extracts the (local Gaussian) auto- and
cross-correlations into the shapes needed in order for the
plot-function to present the result.
}
\keyword{internal}
