% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_CS_input.R
\name{LG_shiny_CS_input}
\alias{LG_shiny_CS_input}
\title{Create \code{radioButtons} and \code{sliderInputs} for the \code{LG_shiny}-interface.}
\usage{
LG_shiny_CS_input(.env, .env2, .part = c("var_local", "p_diag_bw", "levels"))
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}

\item{.part}{One of the values \code{c("var_local", "p_diag_bw",
"levels")}.  This decides which parts of the
interface that should be updated.}
}
\value{
This function creates the \code{radioButtons} and
    \code{sliderInputs} for the \code{LG_shiny}-interface.
}
\description{
This internal function creates the
    \code{LG_shiny}-interface.  It does also deal with details
    related to the internal logs, which enables switching between
    different branch to work better.
}
\keyword{internal}
