% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_explain_interface.R
\name{LG_shiny_explain_interface}
\alias{LG_shiny_explain_interface}
\title{Create the instructions for \code{LG_shiny}.}
\usage{
LG_shiny_explain_interface(input, .env, .env2)
}
\arguments{
\item{input}{A list containing the arguments needed in order to
perform the extraction.}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}

\item{.env2}{Another environment...}
}
\value{
An explanation will be returned.
}
\description{
This helper function for \code{LG_shiny} should return some
explanations that can be added to the interface.  This version is
still in its infancy, i.e. it is mostly a placeholder.
}
\keyword{internal}
