% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_explain_plots.R
\name{LG_shiny_explain_plots}
\alias{LG_shiny_explain_plots}
\title{Explain details about the plots created by \code{LG_shiny}.}
\usage{
LG_shiny_explain_plots(main_dir, input, show_graphical, path_selection_list,
  .env)
}
\arguments{
\item{main_dir}{The part of the \code{data_dir}-argument from the
calling function \code{LG_shiny}, that specifies the main path
to where the information is stored.}

\item{input}{A list containing the arguments needed in order to
perform the extraction.}

\item{show_graphical}{A logical vector that specifies the kind of
data that should be investigated.  Required in order to
identify the correct array to work upon (since more than one
can be available).}

\item{path_selection_list}{A list containing the components needed
in order to identify the correct array to work upon (since more
than one can be available).}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}
}
\value{
An explanation will be returned.
}
\description{
This is a helper function for \code{LG_shiny}, which explains the
content of the plots.  This version is still in its infancy, and is
at the moment mostly intended as a placeholder.
}

