% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_helper.R
\name{LG_shiny_helper}
\alias{LG_shiny_helper}
\title{Deal with the details that concerns the investigation of the data.}
\usage{
LG_shiny_helper(info, input, show_graphical, path_selection_list, default_type,
  .env, .env2)
}
\arguments{
\item{info}{The information object containing all the data about
the content and whereabouts of the required files.}

\item{input}{The list that contains the present status of selected
values in the interactive interface.}

\item{show_graphical}{A logic vector that specifies which kind of
data that should be investigated.}

\item{path_selection_list}{A list containing the components
required to identifying the correct part of \code{info} to work
upon.}

\item{default_type}{Specifies whether "par_five" (default value) or
"par_one" (i.e. Local Gaussian auto-correlation or Local
Gaussian auto-covariance) will be used when both alternatives
are available in the data.}

\item{.env}{An environment needed in order to get hold of the
correct place to read and update some arguments.}

\item{.env2}{Another environment needed in order to get hold of the
correct place to read and update some arguments.}
}
\value{
The result of this function will be that a suitable
    interface is created, and if I can get it to work, then it will
    in addition load the data into the correct environment.
}
\description{
The purpose of this function is to create the selection part of the
interactive interface, i.e. the part that gives selectors that
extracts subsets of the data suited for plotting.  Note that the
loading of the data is taken care of in other functions.
}

