% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_interface_0_RW_log.R
\name{LG_shiny_interface_0_RW_log}
\alias{LG_shiny_interface_0_RW_log}
\title{Link, read, or write, to the deep levels of \code{TS_logging}}
\usage{
LG_shiny_interface_0_RW_log(.env, .env2, .RW = c("L", "R", "W"))
}
\arguments{
\item{.env}{The environment where \code{TS_logging} lives.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.  This argument is only required when we want
to update the \code{input}-list.}

\item{.RW}{Either \code{L} (Link), \code{R} (Read) or \code{W}
(Write).  This argument specifies if we should provide a link
to the environment of interest, should read from it or write to
it.  If left unspecified, the default will be \code{L}.}
}
\value{
When the \code{.RW}-argument is \code{L}, the function will
    return a link to the environment-node containing the values to
    be used when creating the deeper parts of the
    \code{shiny}-interface.  The value \code{R} implies that the
    stored data will be used to update the internal
    \code{input_triggers} and \code{input}-lists, whereas the value
    \code{W} instead writes from these objects to the environment.
}
\description{
Link, read, or write, to the deep levels of \code{TS_logging}
}
\keyword{internal}
