% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_interface_1_TS_info.R
\name{LG_shiny_interface_1_TS_info}
\alias{LG_shiny_interface_1_TS_info}
\title{Create the \code{TS-info}-part of the dynamic \code{LG_shiny}-interface}
\usage{
LG_shiny_interface_1_TS_info(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
The \code{TS_info}-part of the dynamic structure of the
    \code{LG_shiny}-application is created.
}
\description{
This internal function creates the \code{TS-info}-part
    of the dynamic \code{LG_shiny}-interface.
}
\keyword{internal}
