% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_interface_1_TS_info_scribe.R
\name{LG_shiny_interface_1_TS_info_scribe}
\alias{LG_shiny_interface_1_TS_info_scribe}
\title{A scribe for the "old-branch"-parameters from the
\code{LG_shiny}-interface.}
\usage{
LG_shiny_interface_1_TS_info_scribe(.env, .env2)
}
\arguments{
\item{.env}{The environment where the \code{TS_logging}-object
lives.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This helper-function will check if any changes in the top
    level of the interface for the \code{LG_shiny}-function implies
    that an update of stored values should be undertaken.  In
    addition it will perform the initial adjustments of the
    \code{TS_info} part of the \code{input_triggers}, in order for
    the possibility of auto-select to work as intended in the main
    function.
}
\description{
A scribe for the "old-branch"-parameters from the
\code{LG_shiny}-interface.
}
\keyword{internal}
