% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_interface_explanations.R
\name{LG_shiny_interface_explanations}
\alias{LG_shiny_interface_explanations}
\title{Create the part of the \code{LG_shiny}-interface that explains it.}
\usage{
LG_shiny_interface_explanations(.env, .env2, .explain)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}

\item{.explain}{Either \code{interface}, \code{plot} or
\code{show_shiny}, decides which part to be updated.}
}
\value{
This interface-function adds some explanations to the
    \code{LG_shiny}-interface.
}
\description{
Create the part of the \code{LG_shiny}-interface that explains it.
}
\keyword{internal}
