% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_interface_Master.R
\name{LG_shiny_interface_Master}
\alias{LG_shiny_interface_Master}
\title{Master for the creation of the dynamic \code{LG_shiny}-interface.}
\usage{
LG_shiny_interface_Master(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function governs the workers that provides the dynamic
    interface for the \code{LG_shiny}-function.  (The resulting
    updates are stored in the environments \code{.env} and
    \code{.env2}.)
}
\description{
This internal function is the master-function that
    calls different helper-functions required for the dynamic
    \code{LG_shiny}-interface and the associated bookkeeping.
    Depending on the status of the application, it will either
    initiate the log, or it will update the log and allow different
    workers to do the tasks of loading files, fine-tuning different
    parts of the interface, and calling the functions that presents
    the plots and the explanations of the plots.
}
\keyword{internal}
