% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_interface_plots.R
\name{LG_shiny_interface_plots}
\alias{LG_shiny_interface_plots}
\title{Create the plot-related part of the \code{LG_shiny}-interface.}
\usage{
LG_shiny_interface_plots(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function creates the plot (or the code to create it
    outside of the \code{LG_shiny}-application).
}
\description{
Create the plot-related part of the \code{LG_shiny}-interface.
}
\keyword{internal}
