% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_S_input.R
\name{LG_shiny_S_input}
\alias{LG_shiny_S_input}
\title{Create the spectrum-specific part of the \code{LG_shiny}-interface.}
\usage{
LG_shiny_S_input(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function creates the \code{actionButtons} needed for
    the spectrum-specific part of the \code{LG_shiny}-interface.
    This includes the selection of the type of spectrum to
    investigate (auto- or cross-spectrum, with possible
    complex-valued results), and the arguments to be used when
    computing the different spectra (frequency-range, truncation
    point and lag-window-function).
}
\description{
Create the spectrum-specific part of the \code{LG_shiny}-interface.
}
\keyword{internal}
