% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_S_update.R
\name{LG_shiny_S_update}
\alias{LG_shiny_S_update}
\title{Update logged values related to the spectrum-type that is
investigated in the interactive \code{LG_shiny}-interface.}
\usage{
LG_shiny_S_update(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function is a bookkeeping function that deals with
    issues related to the actionButtons that selects the type of
    spectrum to investigate.
}
\description{
Update logged values related to the spectrum-type that is
investigated in the interactive \code{LG_shiny}-interface.
}
\keyword{internal}
