% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_TCS_input.R
\name{LG_shiny_TCS_input}
\alias{LG_shiny_TCS_input}
\title{Create graphical parameters for the \code{LG_shiny}-interface.}
\usage{
LG_shiny_TCS_input(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This interface-function creates another graphical
    parameters part of the \code{LG_shiny}-application.
}
\description{
This internal function will, based on the values in
    the \code{TCS_type} part of the \code{input_triggers}, create
    the next part which is \code{var_local}.  This function will
    also take care of some of the bookkeeping.
}
\keyword{internal}
