% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_writeLines.R
\name{LG_shiny_writeLines}
\alias{LG_shiny_writeLines}
\title{Writes the code needed for the calling of \code{LG_shiny}}
\usage{
LG_shiny_writeLines(main_dir, data_dir, con = stdout())
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{data_dir}{The directory-component of the result from
\code{LG_approx}.}

\item{con}{A connection object or a character string, default value
is \code{stdout()}.  This argument is used when this function
calls \code{writeLines}.}
}
\value{
This function writes a few lines of code that contains the
    details needed in order to start \code{LG_shiny} with the
    arguments \code{main_dir} and \code{data_dir}.  Depending on
    the value of \code{con}, the result will either be sent to the
    terminal or it will be saved to file.
}
\description{
Writes the code needed for the calling of \code{LG_shiny}
}
