% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_spectra.R
\name{LG_spectra}
\alias{LG_spectra}
\title{Compute the local Gaussian spectral densities}
\usage{
LG_spectra(LG_approx_extract_data, omega_vec = NULL,
  omega_length_out = 2^6, window = c("Tukey", "Parzen", "Bartlett"),
  cut_vec = NULL)
}
\arguments{
\item{LG_approx_extract_data}{The list created when
\code{LG_approx_extract} is used upon the Local Gaussian
Approximations, i.e. the restricted subset of data needed for
the computation of the local Gaussian spectra.  The list does
also specify what kind of Local Gaussian Spectral Densities
that should be created i.e. if it should be based on those from
one ore five parameters.}

\item{omega_vec}{A vector of frequencies.  When the default value
\code{NULL} is used, the function will use
\code{omega_length_out} to create the frequency-vector.  Note
that the values given here should be in the interval from
\code{0} to \code{1/2}, since they later on will be multiplied
with \code{2*pi} in the computation.  This makes the baseline
for white noise \code{1} instead of \code{1/2pi}, which is
preferable when creating the plots later on.}

\item{omega_length_out}{The number of values we want to consider in
the interval from 0 to 1/2, that then will be multiplied with
\code{2*pi} in the computation. Reminder: We have an even
function, so the part from -1/2 to 0 are not computed.)  Note
that this argument will be ignored when \code{omega_vec} is
specified.  Default value \code{2^6}.}

\item{window}{The window-function(s) to use on the summands of the
local Gaussian spectra.  The available values are
\code{c("Tukey", "Parzen", "Bartlett")}.  All the window-types
will be computed if no selection is made.  (This might be
modified in a latter incarnation of the code.)}

\item{cut_vec}{The truncation-points to be used for the functions
selected in \code{window}.  If the default value \code{NULL} is
used, the whole range of truncation points from \code{1} to the
highest available (i.e. the highest recorded lag plus one) will
be computed and returned to the work-flow.  A restriction of
\code{cut_vec} to a subset might in particular be a good idea
to do when bootstrap is used to estimate confidence intervals.}
}
\value{
A list containing the information needed for
    \code{LG_shiny} to create plots of the local Gaussian auto- and
    cross-spectra.  The list will contain the components \code{Re}
    and \code{Im} that in essences gives the real and imaginary
    parts of the result.  The component \code{ccr_zero} contains
    the lag-zero contribution for the real part of the cross-
    spectra (the value will always be identical to \code{1} for the
    lag-zero contribution to the auto-spectra, and this is thus not
    stored to reduce the memory load.)
}
\description{
Compute the local Gaussian spectral densities
}
