% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_spectra_scribe.R
\name{LG_spectra_scribe}
\alias{LG_spectra_scribe}
\title{Local Gaussian spectra, scribe-function.}
\usage{
LG_spectra_scribe(main_dir, data_dir, LG_type = c("par_five", "par_one"),
  omega_vec = NULL, omega_length_out = 2^6, window = c("Tukey",
  "Parzen", "Bartlett"), cut_vec = NULL, .method = c("all", "mean",
  "median"))
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{data_dir}{A specification of the directory to be used when
loading and saving data.}

\item{LG_type}{One of \code{c("par_five", "par_one")}, i.e. should the
Local Gaussian Autocorrelations be based on the approach using
five parameters or the simplified approach using only one
parameter.  The default choice is \code{"par_five"}.}

\item{omega_vec}{A vector of frequencies.  When the default value
\code{NULL} is used, the function will use
\code{omega_length_out} to create the frequency-vector.  Note
that the values given here should be in the interval from
\code{0} to \code{1/2}, since they later on will be multiplied
with \code{2*pi} in the computation.  This makes the baseline
for white noise \code{1} instead of \code{1/2pi}, which is
preferable when creating the plots later on.}

\item{omega_length_out}{The number of values we want to consider in
the interval from 0 to 1/2, that then will be multiplied with
\code{2*pi} in the computation. Reminder: We have an even
function, so the part from -1/2 to 0 are not computed.)  Note
that this argument will be ignored when \code{omega_vec} is
specified.  Default value \code{2^6}.}

\item{window}{The window-function(s) to use on the summands of the
local Gaussian spectra.  The available values are
\code{c("Tukey", "Parzen", "Bartlett")}.  All the window-types
will be computed if no selection is made.  (This might be
modified in a latter incarnation of the code.)}

\item{cut_vec}{The truncation-points to be used for the functions
selected in \code{window}.  If the default value \code{NULL} is
used, the whole range of truncation points from \code{1} to the
highest available (i.e. the highest recorded lag plus one) will
be computed and returned to the work-flow.  A restriction of
\code{cut_vec} to a subset might in particular be a good idea
to do when bootstrap is used to estimate confidence intervals.}

\item{.method}{An argument to be used by \code{LG_approx_extract}.
This has to be one or more of the values in \code{c("all",
"mean", "median")}, but the argument will only have an effect
if the length of the \code{content}-dimension in the
\code{LG_type} part of \code{approx_data} is larger than one.
The idea is that this argument will be used with the values
"mean" and "median" while investigating blocks, and the value
"all" when doing an ordinary investigation.}
}
\value{
This function will return a two component list to the
    work-flow.  The first component is the logical value
    \code{done_before} that reveals whether or not the result
    already existed (in the specified file-structure), whereas the
    second component \code{data_dir} gives the location of the
    saved data.
}
\description{
This function takes care of the bookkeeping when the investigation
of a time series by means of local Gaussian spectra is performed.
}
\details{
This function records its arguments and compares them to a
    previously stored information-object, in order to avoid redoing
    previously performed computations.  For new computations,
    relevant data will be extracted from \code{data_dir}, which
    then will be analysed in order to see if memory issues requires
    that the computations should be performed in smaller chunks.
    The function then calls \code{LG_approx_extract} and
    \code{LG_spectra} in order to obtain the result -- and finally
    the result is added to the file-structure while the
    information-object is updated.
}
