% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_spectrum_norm.R
\name{LG_spectrum_norm}
\alias{LG_spectrum_norm}
\title{Compute norms, distances and angles}
\usage{
LG_spectrum_norm(C1_env = NULL, W1 = NULL, C2_env = NULL,
  W2 = NULL)
}
\arguments{
\item{C1_env}{The environment containing the coefficients used for
the first collection of spectra.  These coefficients will be
scaled with the weights from \code{W1}.  Reminder: A copy of the
\code{look_up}-argument must be added to this environment
before this function is called!}

\item{W1}{A vector with the weights to be used for the scaling of
the coefficients given in \code{C1_env}, i.e. the scaling that
truncates the estimates at a given lag. It will always be
assumed that \code{W1} is symmetric around the lag zero
component, so only the positive lags are specified.  It is
furthermore only necessary to specify the nonzero weights, so
this vector can thus be shorter than the one given in
\code{C1_env}.  If this argument is left unspecified, i.e. if
the default value \code{NULL} is given, then it will be assumed
that all the weights should be \code{1}.}

\item{C2_env}{An optional environment that can be used when it is
of interest to compute the distance between two families of
spectra (i.e. the spectra based on coefficients and weights in
\code{C1_env} and \code{W1}, and the spectra based on
\code{C2_env} and \code{W2}).  The default value for
\code{C2_env} is \code{NULL}.  Note that the structure of the
content of this environment does not need to be identical to
the one in \code{C1_env}, since the comparison will be based on
the common components.  Reminder: If the target of interest is
the distance of the spectra from white noise, then it is
sufficient to 'take the square root of norm minus one' in the
auto-correlation case.  For the cross-spectra it is sufficient
to simply take the square root of the norm.  Reminder: A copy
of the \code{look_up}-argument must be added to this
environment before this function is called!}

\item{W2}{An optional vector with the weights to be used for the
scaling of the coefficients given in \code{C2_env}.  The same
assumptions occur for \code{W2} as those specified for
\code{W1}.}
}
\value{
The result of this function is a collection of norms,
    distances and angles related to the investigated spectra.  If
    both \code{C1_env} and \code{C2_env} are given, then the
    following four variations will be returned: C1-local
    vs. C1-global, C2-local vs. C2-global, C1-global vs. C2-global,
    and C1-local vs. C2-local.
}
\description{
Compute norms, distances and angles
}
\keyword{internal}
