% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_splitting.R
\name{LG_splitting}
\alias{LG_splitting}
\title{Split local Gaussian related computations into smaller pieces}
\usage{
LG_splitting(books)
}
\arguments{
\item{books}{The (internal) result we get from
\code{LG_bookkeeping}, that contains all the information needed
for this function.  This includes paths to data-files, the size
and dimension-names of the objects found in these -- and the
arguments that will be used in the desired computation based
upon these data.}
}
\value{
This function will return a "loop-list" to be used
    internally in the \code{LG_scribe}-function.  This is a list
    with the following components: \code{data_files} that gives the
    paths to the files that the main-function must read to get
    access to the data. \code{arg_list} that contains the arguments
    needed for the computation -- and this part will also decide if
    a computation must be split into smaller chunks due to large
    intermediate objects.
}
\description{
To reduce the chance for out of memory problems (see
\code{help("Memory-limits")} for more information), some
computations will be partitioned into smaller chunks.  This
function checks if a splitting is required, and it gives a
partition of the arguments that should (hopefully) ensure that the
available amount of memory is sufficient for the task at hand.
Warning: This is a rather crude approach, that does not measure how
much memory that is available, i.e. it might fail if the work-space
is cluttered with large objects.
}
\details{
The argument and value of this function reflects that it
    is an internal function of the \code{LG_scribe}-functions.
    There are two aspects this function takes care of.  First of
    all it aims at ensuring that the final objects that are saved
    to file(s) stay below some threshold.  Moreover, as some
    intermediate objects can grow very large (in particular when
    dealing with bootstrap-replicates), it will also check if a
    computation should be divided into smaller chunks in order to
    avoid memory problems.  The strategy used in order to obtain
    this is to "split" the array by chopping up some of the
    dimensions.
}
\keyword{internal}
