% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_update_list.R
\name{LG_update_list}
\alias{LG_update_list}
\title{Updating of internal lists.}
\usage{
LG_update_list(.new, .old)
}
\arguments{
\item{.new}{A list that should be used to update (or add) content
to the list given in \code{.old}}

\item{.old}{The list to be updated with the content from the list
given in \code{.new}.}
}
\value{
The result of this function will be an updated version of
    the list given in \code{.old}.  Things in \code{.new} that is
    not present in \code{.old} will be added, whereas those parts
    that occur both places will be updated.
}
\description{
This internal function is primarily used to update the
    \code{curlicues}-list.
}
\keyword{internal}
