% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_Wrapper_Bootstrap.R
\name{LG_Wrapper_Bootstrap}
\alias{LG_Wrapper_Bootstrap}
\title{Wrapper-function for the bootstrap part of local Gaussian
inspection of a time series.}
\usage{
LG_Wrapper_Bootstrap(main_dir, spectra_dir, nb = NULL,
  boot_type = NULL, block_length = NULL, boot_seed = NULL,
  all_statistics = FALSE, log_ = FALSE, lag_max = NULL,
  LG_points = NULL, .bws_mixture = NULL, bw_points = NULL,
  .bws_fixed = NULL, .bws_fixed_only = NULL, content_details = NULL,
  LG_type = NULL, omega_vec = NULL, window = NULL, cut_vec = NULL,
  threshold = 500)
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{spectra_dir}{A vector with names of directories, points to
the local Gaussian spectra, for which we want to find
bootstrap-based confidence intervals.}

\item{nb}{An integer that specifies how many bootstrap-replicates
we want to use in our analysis.  Default value 5 (at least in
the development phase).}

\item{boot_type}{This one can be used to select what kind of
bootstrap-algorithm to be used.  Default value \code{"block"},
but it has not been implemented any other algorithms
yet... (Reminder: It would be be nice to include the
'network-duality' approach in this context -- but then I fear I
might need to implement that routine from scratch.)}

\item{block_length}{The length of the blocks to be used when
\code{boot_type="block"} is used.  Default value 20 in the
development phase, but I suppose in general it should be some
formula based on the time series under investigation.}

\item{boot_seed}{Use this to enable reproducible results.  The
default value \code{NULL} will trigger a random seed to be
selected for this value (that then will be recorded in case a
reproduction of the result is desired later on).}

\item{all_statistics}{A logical value, default \code{FALSE}.  The
default will return a bunch of quantiles based on the
collection of bootstrapped values, which can be used when we
need estimates for the confidence intervals.
\code{all_statistics} equals \code{TRUE} will in addition
return a bunch of statistics related to the distribution of the
bootstrapped values.}

\item{log_}{A logic argument with default value \code{FALSE}.  This
decides if the statistics should be computed based on the
logarithmic values of our replicates (when that makes sense,
i.e. when we have nonzero values).  It might be preferable to
use \code{log_=TRUE}, but as we have no guarantee that it
occasionally might not occur negative values for the estimated
local Gaussian spectra, the default has nevertheless been set
to \code{FALSE}.}

\item{lag_max}{The maximal number of lags we want to include in our
analysis.  This number will also affect the
\code{adjustment}-attribute on the bootstrapped time series.}

\item{LG_points}{An array that specifies the point at which it is
desired to compute the local Gaussian estimates.  The default
value \code{NULL} will imply that the values used in the
computations upon the original data are recycled.  Values can
alternatively be computed by the help of the function
\code{LG_select_points}, but these must then be a subset of the
original values.}

\item{.bws_mixture}{One (or more) of the values in
\code{c("mixture", "local", "global")}.  The alternatives
\code{local} and \code{global} will respectively only select
the nearest neighbour or global, whereas \code{mixture} selects
the largest of the two alternatives.}

\item{bw_points}{The percentage(s) of the observations that we want
inside the "bandwidth-square".  If \code{.bws_mixture} is
selected to be \code{global}, then this argument will be
ignored. and no nearest neighbours will be computed.}

\item{.bws_fixed}{A vector of non-negative real values, that can be
used to specify fixed values for the bandwidths (which might be
of interest to do in a preliminary analysis).}

\item{.bws_fixed_only}{A logic value that can be used to drop the
rather time-consuming data-driven estimation of bandwidths in
favour of the simplified approach where fixed bandwidths are
used instead.  Note that \code{.bws_fixed} must be specified
when \code{.bws_fixed_only} are set to \code{TRUE}.}

\item{content_details}{A value from \code{c("rho_only",
"rho_log.fun", "rho_all")}, which decides the amount of details
stored from the local Gaussian approximations.}

\item{LG_type}{One of \code{c("par_five", "par_one")}, i.e. should
the Local Gaussian Autocorrelations be based on the approach
using five parameters or the simplified approach using only one
parameter.  When \code{NULL} is used, the function will use the
\code{LG_type}-argument from the original local Gaussian
spectral analysis.}

\item{omega_vec}{A vector of frequencies.  This must be a subset of
the values used when analysing the original time series.  If
the default value \code{NULL} is used, then the function will
look up the \code{omega_vec}-argument from the original local
Gaussian spectral analysis.}

\item{window}{The window-function(s) to use on the summands of the
local Gaussian spectra. This must be a subset of the values
used when analysing the original time series.  If the default
value \code{NULL} is used, then the function will look up the
\code{window}-argument from the original local
Gaussian spectral analysis.}

\item{cut_vec}{The truncation-points to be used for the functions
selected in \code{window}.  This must be a subset of the values
used when analysing the original time series.  If the default
value \code{NULL} is used, then the function will look up the
\code{cut_vec}-argument from the original local Gaussian
spectral analysis.  A restriction of \code{cut_vec} to a
relevant subset (based on an inspection of the original result)
might be a good idea, since the computational time then can be
reduced to a fraction.}

\item{threshold}{An integer, default value 500 (measured in MB),
that tells the program when a computation should be divided
into smaller chunks.  This reduces the chance of memory-related
problems, but as the present incarnation of \code{LG_splitter}
are rather stupid such problems might still occur for long
time-series where a huge number of lags are included.}
}
\value{
This function will compute local Gaussian approximations
    for all the bootstrap-replicates of the original time series
    \code{TS}, and then it will compute the local Gaussian spectra
    and extract information about them in accordance with
    \code{all_statistics}.  The results of these computations will
    be saved in the file-hierarchy and the info-file will be
    updated with details about the arguments.  The latter part is
    done in order to keep track of previous computations, such that
    we avoid redoing stuff that has been performed before.  Details
    about the saved files are returned to the work-flow, in a list
    with the two components \code{approx_note} and
    \code{spectra_note}.
}
\description{
Wrapper-function for the bootstrap part of local Gaussian
inspection of a time series.
}
