% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_Wrapper_Original.R
\name{LG_Wrapper_Original}
\alias{LG_Wrapper_Original}
\title{Wrapper-function for local Gaussian inspection of a time series}
\usage{
LG_Wrapper_Original(main_dir, data_dir, TS, lag_max = ceiling(3 *
  sqrt(length(TS))), LG_points, .bws_mixture = c("mixture", "local",
  "global"), bw_points = c(25, 35), .bws_fixed = NULL,
  .bws_fixed_only = FALSE, omega_vec = NULL, omega_length_out = 2^6,
  window = c("Tukey", "Parzen", "Bartlett"), LG_type = c("par_five",
  "par_one"), cut_vec = NULL)
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{data_dir}{A specification of the directory to be used when
loading and saving data.}

\item{TS}{The time series we want to investigate by means of
techniques based on local Gaussian approximations.  Note that
this argument in the "wrapper"- and "scribe"-functions can be
given as a character vector that specifies the file where
\code{TS} from \code{TS_LG_object} was stored.}

\item{lag_max}{The number of lags to include in the analysis.  The
default value \code{ceiling(3*sqrt(length(TS)))} will probably
in most cases include more lags than actually required.  It
might thus be worthwhile to run some test and see if a smaller
value can be used instead.}

\item{LG_points}{An array that specifies the point at which it is
desired to compute the local Gaussian estimates.  Use the
function \code{LG_select_points} to create this argument.}

\item{.bws_mixture}{An argument that specifies how the global
bandwidths and those obtained by the nearest-neighbour strategy
should be combined.  The three available options are
\code{c("mixture", "local", "global")}, which have the
following effects.  The alternatives \code{local} and
\code{global} will respectively only select the nearest
neighbour or global.  These alternatives seems however to not
work well when used on some parts of the lagged pairs of
interest, i.e. the nearest neighbour values might be to "small"
in the center of the distribution, whereas the global
bandwidths seems to fail in the periphery of the distribution.
The alternative \code{mixture} attempt to resolve this by (for
each grid point) selecting the largest of the two alternative
bandwidths.  Note that the value of \code{.bws_mixture} decides
how much information that is computed, i.e. the alternative
\code{local} will turn off the computation of global
bandwidths.  However, the computations of the nearest neighbour
bandwidths will also be computed when the alternative
\code{global} is used, since it does not take long to compute
and it is that function that creates the array we need as a
mould for the result.  If the user does not make a selection,
then all three alternatives will be computed.}

\item{bw_points}{A vector, default \code{c(25, 35)}, that specifies
the percentage of the observations that we want inside the
"bandwidth-square".  If \code{.bws_mixture} is selected to be
\code{global}, then this argument will be ignored. and no
nearest neighbours will be computed.}

\item{.bws_fixed}{A vector of non-negative real values, that can be
used to specify fixed values for the bandwidths (which might be
of interest to do in a preliminary analysis).  The default
value \code{NULL} will prevent the computation of Local
Gaussian Estimates based on fixed bandwidths.}

\item{.bws_fixed_only}{A logic value, default \code{FALSE}, that
can be used to drop the rather time-consuming data-driven
estimation of bandwidths in favour of the simplified approach
where fixed bandwidths are used instead.  Note that
\code{.bws_fixed} must be specified when \code{.bws_fixed_only}
are set to \code{TRUE}.}

\item{omega_vec}{A vector of frequencies.  When the default value
\code{NULL} is used, the function will use
\code{omega_length_out} to create the frequency-vector.  Note
that the values given here should be in the interval from
\code{0} to \code{1/2}, since they later on will be multiplied
with \code{2*pi} in the computation.  This makes the baseline
for white noise \code{1} instead of \code{1/2pi}, which is
preferable when creating the plots later on.}

\item{omega_length_out}{The number of values we want to consider in
the interval from 0 to 1/2, that then will be multiplied with
\code{2*pi} in the computation. Reminder: We have an even
function, so the part from -1/2 to 0 are not computed.)  Note
that this argument will be ignored when \code{omega_vec} is
specified.  Default value \code{2^6}.}

\item{window}{The window-function(s) to use on the summands of the
local Gaussian spectra.  The available values are
\code{c("Tukey", "Parzen", "Bartlett")}.  All the window-types
will be computed if no selection is made.  (This might be
modified in a latter incarnation of the code.)}

\item{LG_type}{One of \code{c("par_five", "par_one")}, i.e. should the
Local Gaussian Autocorrelations be based on the approach using
five parameters or the simplified approach using only one
parameter.  The default choice is \code{"par_five"}.}

\item{cut_vec}{The truncation-points to be used for the functions
selected in \code{window}.  If the default value \code{NULL} is
used, the whole range of truncation points from \code{1} to the
highest available (i.e. the highest recorded lag plus one) will
be computed and returned to the work-flow.  A restriction of
\code{cut_vec} to a subset might in particular be a good idea
to do when bootstrap is used to estimate confidence intervals.}
}
\value{
This function will compute local Gaussian approximations
    and local Gaussian spectra in accordance with the given
    arguments.  The results of these computations will be saved in
    the file-hierarchy and the info-file will be updated with
    details about the arguments.  The latter part is done in order
    to keep track of previous computations, such that we avoid
    redoing stuff that has been performed before (in the present
    \code{data_dir}).  Details about the saved files are returned
    to the work-flow, in a list with the two components
    \code{approx_note} and \code{spectra_note}.
}
\description{
Wrapper-function for local Gaussian inspection of a time series
}
\details{
This function will take care of the pesky details when it
    comes to calling the functions that based on a time series will
    analyse it by means of local Gaussian techniques, i.e. by
    computing a bunch of local Gaussian approximations for the lags
    and then use these to create local Gaussian spectra.
}

