% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgaussSpec-package.r
\docType{package}
\name{localgaussSpec}
\alias{localgaussSpec}
\alias{localgaussSpec-package}
\title{localgaussSpec}
\description{
A package for the investigation of time series by means of
Local Gaussian Spectral Densities.
}
\details{
The localgaussSpec package contains the functions needed for the
investigation by means of local Gaussian approximations.  Moreover,
it also contains copies of the scripts needed for a reproduction of
results mentioned in ...  See <helper-function> for details.
}
\section{Simulating data}{

Some stuff about the functions used for simulations, and the
format of the result.
}

\section{Computation of local Gaussian approximations}{

blah blah...
}

\section{Computation of local Gaussian spectra}{

blah blah...
}

\section{Computation of bootstrap-based confidence intervals}{

blah blah...
}

