% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{myTS}
\alias{myTS}
\title{My Time Series}
\usage{
myTS(TS_key, ...)
}
\arguments{
\item{TS_key}{character, key that selects one of the time series
from \code{TS_families}}

\item{...}{dotsMethods (i.e. optional), any arguments to be
delivered to the function selected by \code{TS_key}.}
}
\value{
\code{fun}, a function that can generate the desired time
    series.  Note that the formals of function selected by
    \code{TS_key} will have been updated with any the parameters
    given via the \code{dotsMethods}-argument.

\code{args.i}, a list of the (updated) default values that
    are used by the function in \code{fun}.

\code{args.f}, a character-string of the default values,
    that can be used in the file-name for the data-file generated
    from the time series.
}
\description{
My Time Series
}
\details{
Given a key for a time series in \code{TS_families}, and
    relevant parameters for it (delivered by the
    dotsMethods-argument \code{...}), this function will return the
    function that will be used in \code{myTS.sample} to generate
    the desired sample from the specified time series.
}
\keyword{internal}
\keyword{series}
\keyword{time}
