% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\docType{data}
\name{myWindows}
\alias{myWindows}
\title{A list with the lag-window functions that is used to smooth the
estimated spectral densities.}
\format{An object of class \code{list} of length 3.}
\usage{
myWindows
}
\description{
At the moment only three alternatives are present: "Tukey",
"Parzen" and "Bartlett".
}
\details{
The \code{.cut}-argument that is common for all of the lag-window
functions is the first lag that is given the weight zero.  Note
that the truncation level \code{m} that is used for the truncated
spectra refers to the last non-zero lag, which implies that the
code later on must adjust for this difference.
}
\keyword{internal}
