% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_adjust_gamma_rho.R
\name{TS_adjust_gamma_rho}
\alias{TS_adjust_gamma_rho}
\title{Adjustments for \eqn{$\gamma$} and $\eqn{$\rho$}.}
\usage{
TS_adjust_gamma_rho(TS, lag_max = NULL, type = c("rho", "gamma"),
  .adjustment_rule = "data", all_details = FALSE)
}
\arguments{
\item{TS}{The time series object that we want to work upon.  It's
assumed that this has been converted into an array, with one
dimension-name equal to "content".}

\item{lag_max}{Integer that decides the number of scaling and
adjustment-terms that will be returned.  The default value
\code{NULL} will imply that these terms are included from lag 0
to \code{length(TS)-2}.  (In practise this will be way above
what will be used in the estimates, so this argument should be
specified in order to avoid wasting computational resources.)}

\item{type}{One of \code{c("rho", "gamma")}, i.e. should
the adjustment-terms be computed for data based on local
Gaussian correlations or local Gaussian covariances.  If no
selections are made, both alternatives will be computed.}

\item{.adjustment_rule}{Either a non-negative number, or "data".
This will be added as an attribute to the normalised time
series, and later on it will decide if any finite-sample
adjustment should be used for the estimated local Gaussian
autocorrelations.  Note that no adjustments will be performed
when \code{.adjustment_rule=0}.}

\item{all_details}{Logic value, default \code{FALSE}, that
influences the amount of information returned to the workflow.}
}
\value{
An array will be returned giving the scaling and adjustment
    terms required to go from estimates covariances and
    correlations to the estimates of autocovariances and
    autocorrelations.  If \code{all_details} is \code{TRUE}, then
    some intermediate estimates will be included in the result.
}
\description{
The estimation formulas used for autocovariances and
autocorrelations at lag \eqn{$h$} are connected to the estimation
formulas of the covariances and correlation of the corresponding
lagged pairs.  This function computes the scaling and adjustments
components that we need in other functions later on.
}

