% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_boot_sample.R
\name{TS_boot_sample}
\alias{TS_boot_sample}
\title{Create bootstrap-replicates of a time series.}
\usage{
TS_boot_sample(
  TS,
  main_dir,
  save_dir,
  nb = 5,
  boot_type = c("cibbb_tuples", "block"),
  block_length = 20,
  boot_seed = NULL
)
}
\arguments{
\item{TS}{The original time series that we want to create
bootstrap-replicates of.}

\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{save_dir}{A specification of the directory to be used when
saving (and loading) data.  The default value \code{NULL} will
turn of the interaction with disk.}

\item{nb}{An integer that specifies how many bootstrap-replicates
we want to use in our analysis.  Default value 5 (at least in
the development phase).}

\item{boot_type}{This argument should be either
\code{"cibbb_tuples"} or \code{"block"}, the former gives an
implementation of the circular index-based block bootstrap for
tuples adjustment, whereas the latter gives the ordinary block
bootstrap.  The default option is \code{"cibbb_tuples"}.}

\item{block_length}{The length of the blocks to be used when
\code{boot_type="block"} is used.  Default value 20 in the
development phase, but I suppose in general it should be some
formula based on the time series under investigation.}

\item{boot_seed}{Use this to enable reproducible results.  The
default value \code{NULL} will trigger a random seed to be
selected for this value (that then will be recorded in case a
reproduction of the result is desired later on).}
}
\value{
A list with the following three components is returned to
    the internal workflow.

\describe{

\item{main_dir}{The path (to the hierarchy) given by the
    \code{main_dir}-argument.}

\item{TS}{The internal (in-hierarchy) path to the saved data.}

\item{save_dir}{The (in-hierarchy) name of the directory that the
    files are saved into.}

}
}
\description{
This internal function will, based on the specified
    bootstrapping-algorithm and the given time series, create and
    save a matrix of bootstrapped time series.
}
\details{
The bootstrapped replicates are stored in a matrix with
    \code{nb} rows and \code{length(TS)} columns, i.e.  each row is
    a bootstrapped replicate of \code{TS}.
}
\keyword{internal}
