% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_CI.R
\name{TS_CI}
\alias{TS_CI}
\title{Compute confidence intervals based on spectral estimates}
\usage{
TS_CI(main_dir, spectral_content, all_statistics, log_, ...)
}
\arguments{
\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{spectral_content}{The path needed to open the file containing
the estimated spectral densities.}

\item{all_statistics}{A logical value, default \code{FALSE}.  The
default will return a bunch of quantiles based on the
collection of bootstrapped values, which can be used when we
need estimates for the confidence intervals.
\code{all_statistics} equals \code{TRUE} will in addition
return a bunch of statistics related to the distribution of the
bootstrapped values.}

\item{log_}{A logic argument with default value \code{FALSE}.  This
decides if the statistics should be computed based on the
logarithmic values of our replicates (when that makes sense,
i.e. when we have nonzero values).  It might be preferable to
use \code{log_=TRUE}, but as we have no guarantee that it
occasionally might not occur negative values for the estimated
local Gaussian spectra, the default has nevertheless been set
to \code{FALSE}.}

\item{...}{Additional arguments to be passed to the
apply-functions.}
}
\value{
A file will be saved to disk, and the function will return
    information to be used when updating the corresponding
    \code{info}-object that is maintaned by the calling function.
}
\description{
This is a helper-function to deal with the computation of the
confidence intervals for the global spectral densities.  This is
done in order to create a global reference which the local Gaussian
spectral densities can be compared against.
}
