% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_cibb_block.R
\name{TS_cibb_block}
\alias{TS_cibb_block}
\title{Circular index-based block bootstrap for tuples}
\usage{
TS_cibb_block(.indices, .lag)
}
\arguments{
\item{.indices}{The initial vector of (resampled) indices.}

\item{.lag}{The lag-value that should be used.}
}
\value{
This helper function returns a list with two vectors, one
    named \code{first} and one named \code{second}.  These are the
    indices needed for the circular index-based block bootstrap for
    tuples.
}
\description{
Circular index-based block bootstrap for tuples
}
\keyword{internal}
