% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localgaussSpec-package.r
\docType{data}
\name{TS_families}
\alias{TS_families}
\title{Time Series Families}
\format{A list of lists, the names at the top-level will be used as
    "keys" that selects the relevant sub-list.  The sublists
    contanins the parts \code{fun}, \code{package}, \code{args},
    \code{size_name} and \code{method}, as described below:

\describe{

\item{fun}{The name of the desired function, as a
    character-string.}

\item{package}{The name of the package that contains \code{fun}, as
    a character-string.}

\item{args}{A list containing the default arguments to be used in
    \code{fun}.  Reminder: In order for the package to compile, it
    might be necessary to quote the list if there are references to
    packages not imported into this package.}

\item{size_name}{The name of the argument of \code{fun} that
    specifies the number of samples to be created.}

\item{method}{A character string that has to be specified if
    \code{fun} needs to be "extracted" from some deeper level of
    \code{package}.  Simply drop this part when it is superfluous.}

}}
\usage{
TS_families
}
\description{
The function that computes the samples will use keywords to select
a relevant part from the list \code{TS_families}.  The idea is that
this should help document all the parameters used in the sampling
procedure, such that proper calls can be stored that later on can
be resurrected if a reproduction of the computation is desired.
}
\keyword{internal}
