% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_key_dmt_bivariate.R
\name{TS_key_dmt_bivariate}
\alias{TS_key_dmt_bivariate}
\title{"Local trigonometric" bivariate samples based on cosine-functions}
\usage{
TS_key_dmt_bivariate(n, first_dmt, phase_adjustment = 0)
}
\arguments{
\item{n}{The number of samples to create.}

\item{first_dmt}{A list with all the arguments for the
\code{TS_key_dmt}-function.}

\item{phase_adjustment}{The numerical value(s) given here will be
added to the \code{theta}-value(s) in the
\code{first_dmt}-value.  A single number will imply that all
the cosine-functions will have the same phase-adjustment,
whereas different numbers for each \code{theta}-value in
\code{first_dmt} will give individual phase-adjustments.}
}
\value{
A bivariate "local trigonometric" sample is returned.
}
\description{
The \code{TS_key_dmt_bivariate}-function is based on a
    a simple wrapper of the \code{TS_key_dmt}-function, which
    enables a bivariate "local trigonometric" example to be
    generated.  The basic idea is that two components are computed
    in succession by the help of \code{TS_key_dmt}, and that the
    arguments of the second component only differs by a
    phase-adjustment of the \code{theta}-argument.
}
\note{
This internal function has been added to the
    \code{TS_families}-list, and its arguments should thus be given
    to \code{TS_sample}, which will send them to this function by
    the internal helper-function \code{TS_sample_helper}.
}
