% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_key_dmt_rmgarch.R
\name{TS_key_rmgarch}
\alias{TS_key_rmgarch}
\title{A simple multivariate GARCH-models from the \code{rmgarch}-package.}
\usage{
TS_key_rmgarch(n.sim = 100, m.sim = 10, uspec = NULL, data = NULL,
  rseed = NULL)
}
\arguments{
\item{n.sim}{The simulation horizon.  Default value 100.}

\item{m.sim}{The number of simulations.  Default value 10.}

\item{uspec}{The specification of the GARCH-model for the
univariate margins. Default value \code{NULL}.  If this is not
specified, then \code{ugarchspec} from the
\code{rugarch}-package will be used to specify a trivial model.}

\item{data}{The multivariate time series sample.  The parameters of
interest for the multivariate GARCH-model will be extracted
from the model fitted to this sample.}

\item{rseed}{Optional seeding value(s) for the random number
generator.  This should be of length equal to m.sim.}
}
\value{
A multivariate GARCH-type time series is returned.
}
\description{
The \code{TS_key_rmgarch}-function is in essence a
    wrapper around the \code{cgarchsim}-function from the
    \code{rmgarch}-package, i.e. this is based on the
    "Copula GARCH simulation".  The key idea is that the user can
    feed this wrapper the multivariate sample of interest (the
    \code{data}-argument), and a specification of the univariate
    GARCH-model, and then this function will take care of the rest.
}
\note{
This internal function has been added to the
    \code{TS_families}-list, and its arguments should thus be given
    to \code{TS_sample}, which will send them to this function by
    the internal helper-function \code{TS_sample_helper}.

The \code{rmgarch}-package also contains other simulation
    alternatives, but these have not been implemented.  These
    alternatives are 'dccsim' (Dynamical Conditional Correlation)
    and 'gogarchsim' (Generalised Orthogonal).
}
\keyword{internal}
