% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_families.R
\name{TS_key_rnorm}
\alias{TS_key_rnorm}
\title{The univariate Gaussian case for \code{TS_key}}
\description{
The \code{TS_key_rnorm}-case part of \code{TS_families}
    is simply an implementation of \code{rmnorm} from the
    \code{stats}-package.
}
\note{
The arguments \code{mean} and \code{sd} should be given to
    \code{TS_sample}, which will send them to this function by the
    internal helper-function \code{TS_sample_helper}.
}
