% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_LG_normalisation.R
\name{TS_LG_normalisation}
\alias{TS_LG_normalisation}
\title{Create a normalised time series from the original ones.}
\usage{
TS_LG_normalisation(TS, .remove_ties = TRUE)
}
\arguments{
\item{TS}{The time series object to normalise.  It's assumed that
this is structured in an array, with one dimension-name being
"content".}

\item{.remove_ties}{A logical value, default \code{TRUE}, in which
case the presence of ties will trigger a minor perturbation of
the data.  Note: Whenever this happens, \code{set.seed(1)}
will be used in order to ensure reproducibility.}
}
\value{
A normalised version of the time series of interest.
}
\description{
Create a normalised time series from the original ones.
}
\details{
A normalisation of the time series under consideration is
    preferable when an analysis based on the local Gaussian
    approximations are of interest.  The intention of this function
    is to allow for easier adjustment of the normalisation regime,
    i.e if the normalisation should be based on the cumulative
    density function (or other options that might be added later
    on).  This function will also add some attributes to the
    result, with the required instruction for the finite-sample
    adjustments to be used when computing the local Gaussian
    autocorrelations.
}
\keyword{internal}
