% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_LG_object.R
\name{TS_LG_object}
\alias{TS_LG_object}
\title{Prepare a time series for a local Gaussian inspection}
\usage{
TS_LG_object(TS_data, details = NULL, main_dir = "~/LG_DATA",
  save_dir = NULL, .normalisation_rule = c("ecdf", "logspline"),
  .adjustment_rule = "data", .remove_ties = TRUE)
}
\arguments{
\item{TS_data}{The time series data we will work upon.  This can
either be an observed or a simulated time series.  If
\code{TS_data} are of class "TS_simulated", i.e. it has been
generated by 'TS_sample', then the information stored in it
will be used to create \code{save_dir}.  Note that
\code{TS_data} can be univariate or multivariate.  A
univariate time series can be given as a vector, whereas a
multivariate must have the observations along the rows and the
variates along the columns.  (The program terminates if the
number of rows are lower than the number of columns.)}

\item{details}{This can be used to add a reminder that will be
shown under the interactive investigation later on.  The
default value \code{NULL} will imply that no information is
shown.}

\item{main_dir}{The main directory into which the information will
be stored.  Default value \code{"~/LG_DATA"}, i.e. a specially
designed directory in the home directory of your file-system.
If the proposed default directory does not exists, then
\code{TS_LG_object} will create it, but otherwise it's a
requirement that only existing directories can be used.  This
is done as a precaution against accidentally ending up with
unintended data-directories all over the file-system.}

\item{save_dir}{The sub-directory of \code{main_dir} where all the
stuff related to \code{TS_data} will be saved.  Default value
\code{NULL}, but with the following defaults in the code for
what to replace it with: When \code{TS_data} has been created
by \code{TS_sample}, a value for \code{save_dir} will be
created from the information in \code{TS_data}, and any attempt
from the user to create another name will be outright ignored.
If no value is given for \code{save_dir} (and none can be
computed from \code{TS_data}), then the default value from
\code{LG_default} will be used to create \code{save_dir}.  If the
user specifies \code{save_dir} (for a time-series not
originating from \code{TS_sample}), then that name will be
used, but only if no previous directories happens to have that
name too -- if that should be the case, the program will
terminate and inform the user about it.}

\item{.normalisation_rule}{A value from \code{c("ecdf",
"logspline")} that decides which kind of normalisation rule to
use upon the data.  The "logspline"-alternative seems to give a
more faithful representation of the data, but can encounter
problems if there's "to much data" in some regions (which can
occur for a long time series if some really extreme
observations are present).  The "ecdf" doesn't encounter such
problems, and it will also be much quicker to compute, which
for larger time series might be an important aspect to keep in
mind.}

\item{.adjustment_rule}{Either a non-negative number, or "data".
This will be added as an attribute to the result, and later on
it will decide if any finite-sample adjustment should be used
for the estimated local Gaussian autocorrelations.  Note that
no adjustments will be performed when
\code{.adjustment_rule=0}.}

\item{.remove_ties}{A logical value, default \code{TRUE}, in which
case the presence of ties will trigger a minor perturbation of
the data.  Note: Whenever this happens, \code{set.seed(1)}
will be used in order to ensure reproducibility.}
}
\value{
This function will take care of some file-handling before
    it returns a two-component list to the work-flow.  The first
    component \code{TS_done_before} is a logic value to inform the
    use whether or not the time series from \code{TS_data} already
    was stored in the folder \code{main_dir}.  The second
    component, \code{result}, is a list whose format depends upon
    whether or not \code{TS_data} was created by \code{TS_sample}
    -- and some of the content are only connected to the internal
    work-flow of this function.  The four parts of \code{result}
    that always is present is \code{TS_key} (the origin of the time
    series), \code{TS} (the values), \code{N} (the number of
    observations), and \code{save_dir} (the path to the
    save-directory).  These four values will be used by the
    functions that analyses \code{TS} based upon Local Gaussian
    Approximations and Local Gaussian Spectral Densities.
}
\description{
Prepare a time series for a local Gaussian inspection
}
\details{
This function will for a given (sample from a) time series
    create a directory that will be used to store all the files
    that occur during the local Gaussian analysis.  The function
    will in addition create a local info-file to take care of the
    subsequent bookkeeping.  Moreover, this function will also
    maintain a global info-file in the \code{main_dir}-directory,
    where information about the top-level part of the
    directory-structure is stored -- in order to avoid the
    despicable situation that more than one directory stores
    information about the exact same set of data.  The sample will
    be nudged a tiny bit (always using 1 as the seed-value) if ties
    are detected in it.  The algorithm ensures that this nudge is
    of a small order compared to the original values.
}

