% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_sample.R
\name{TS_sample}
\alias{TS_sample}
\title{Time Series Sample-Generator}
\usage{
TS_sample(
  TS_key = "rnorm",
  N = 500,
  nr_samples = 1,
  ...,
  .seed = NULL,
  .kind_vstr_list = NULL
)
}
\arguments{
\item{TS_key}{A key, i.e. a \code{character}, corresponding to an
element in \code{TS_families}.  Default value \code{rnorm},
i.e. "White Noise Gaussian".}

\item{N}{The desired length of the time series.  Default value
\code{500}.}

\item{nr_samples}{The desired number of independent samples to be
produced.  The default value \code{1} gives a single sample,
which might be preferable for an initial investigation that
includes many different bandwidths and looks at a wide range of
points.  For pointwise confidence bands it might be preferable
to use at least 100 for this argument.}

\item{...}{\code{dotsMethods}-strategy for feeding parameters to
the function (identified by \code{TS_key}) that generates the
time series.}

\item{.seed}{Use this to enable reproducible results.  Default value
\code{NULL} (it will be generated and recorded in the code).}

\item{.kind_vstr_list}{This can be used to create a list with the
values for \code{kind}, \code{normal.kind} and \code{vstr}.
(See the help-page of \code{Random} for details about these
three arguments.)  Note that the default value \code{NULL} will
imply that the function \code{set_seed} will be used to create
the required list based on the present settings.}
}
\value{
This function returns a list with the following four parts:

\describe{

\item{TS}{A 3-dimensional array with the resulting time series,
    generated according to the specified arguments.  The array
    contains three dimensions in order to have a unified solution
    for univariate and multivariate solutions.  The first dimension
    reflects the number of samples, as given by \code{nr_samples}.
    The second dimension reflects the length of the samples, as
    given by \code{N}.  The third dimension reflects the number of
    variables in the time series, which is decided by \code{TS_key}
    and the arguments given to \code{...}}

\item{TS_data}{Additional details used by \code{TS_LG_object} when
    the data is saved to disk.}

\item{spy_report}{An environment that contains all the arguments
    that was used for the present computation.}

\item{seed_vec}{This part reflects that the internal workings of
    the code creates a vector of seeds (based on \code{.seed}) that
    then can be used to reproduce an individual series later on.
    This is of course only of interest when \code{nr_samples} is
    larger than one, and we for some reason later on would like to
    do an in depth analysis of one of the resulting time series.}

}
}
\description{
Time Series Sample-Generator
}
\details{
This function will create one or more time series based on
    the "keys" stored in \code{TS_families}, with emphasis on also
    storing the required arguments needed in order to reproduce it
    later on.
}
