% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_sample_helper.R
\name{TS_sample_helper}
\alias{TS_sample_helper}
\title{Helper-function for \code{TS_sample}}
\usage{
TS_sample_helper(TS_key, ...)
}
\arguments{
\item{TS_key}{character, key that selects one of the time series
from \code{TS_families}}

\item{...}{dotsMethods (i.e. optional), any arguments to be
delivered to the function selected by \code{TS_key}.}
}
\value{
This function returns a list with the following five nodes
    to the internal workflow of \code{TS_sample}:

\describe{

\item{fun}{A local copy of the desired function, which can generate
    the desired time series.  Note that the formals of the function
    selected by \code{TS_key} will have been updated with the
    parameters given via the \code{dotsMethods}-argument.}

\item{args.i}{A list with the (updated) default values that
    are used by the function in \code{fun}.}

\item{args.f}{A character-string of the given values.}

\item{TS_key}{The \code{TS_key}-value given to this function.}

\item{fun_data}{An updated version of the \code{TS_key}-sublist of
    \code{TS_families}, in which the arguments have been updated
    based on the values given to the \code{...}-argument.}

}
}
\description{
This internal function is a helper-function for
    \code{TS_sample}, which based on the \code{TS_key} and the
    specified arguments performs an initial sanity test and then
    returns the function needed for the desired simulation.
}
\details{
Given a \code{TS_key} (the name of a sublist of
    \code{TS_families}), and relevant parameters for it (delivered
    by the dotsMethods-argument \code{...}), this function will
    return the function that will be used in \code{TS_sample} to
    generate the desired sample from the specified time series.
}
\keyword{internal}
\keyword{series}
\keyword{time}
