% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_spectral.R
\name{TS_spectral}
\alias{TS_spectral}
\title{Compute old-fashioned spectral estimates}
\usage{
TS_spectral(acr_data, omega_vec, window = c("Tukey", "Parzen",
  "Bartlett"), cut_vec = NULL, .method = c("all", "mean", "median"))
}
\arguments{
\item{acr_data}{A list giving the information required in order to
read the estimated autocorrelations from file, and moreover
compute the path at which the estimated spectral densities
should be saved.}

\item{omega_vec}{A vector of frequencies}

\item{window}{The window-function(s) to be used.  The available
values are \code{c("Tukey", "Parzen", "Bartlett")}.  All the
window-types will be computed if no selection is made.  (This
might be modified in a latter incarnation of the code.)}

\item{cut_vec}{The truncation-points to be used for the functions
selected in \code{window}.  If the default value \code{NULL} is
used, the whole range of truncation points from \code{1} to the
highest available (i.e. the highest recorded lag plus one) will
be computed and returned to the work-flow.  A restriction of
\code{cut_vec} to a subset might in particular be a good idea
to do when bootstrap is used to estimate confidence intervals.}

\item{.method}{One or more of the values in \code{c("all", "mean",
"median")}, but the argument will only have an effect if the
length of the \code{content}-dimension in the \code{LG_type}
part of \code{approx_data} is larger than one.  The idea is
that this argument will be used with the values "mean" and
"median" while investigating blocks, and the value "all" when
doing a bootstrap-investigation.}
}
\value{
A file will be saved to disk, and the function will return
    information to be used when updating the corresponding
    \code{info}-object that is maintaned by the calling function.
}
\description{
This function computes estimates of the spectral densities for a
time series the "old-fashioned" way, i.e. using window-functions in
order to smooth the estimates.  This is done in order to create a
global base-functions, against which the local Gaussian spectral
densities can be compared.  Note that this functions is kind of a
wrapper around \code{LG_spectra}, where the core of the computation
is performed.
}
