% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_bookkeeping.R
\name{LG_bookkeeping}
\alias{LG_bookkeeping}
\title{Keep the records up to date}
\usage{
LG_bookkeeping(spy_report)
}
\arguments{
\item{spy_report}{The result returned by \code{spy} when used on
the targeted function.}
}
\value{
The result will be a list with either three or six
    components.  The first component, the logic value
    \code{done_before}, tells us if there already has been
    performed (and stored in \code{data_dir} an identical
    computation.  If this is the case, then the second component of
    the result, also named \code{data_dir}, will tell us the
    directory to find the result -- whereas the third component
    \code{data_files} gives the files containing the results.  If
    no similar computation has been found, then \code{data_dir} and
    \code{data_files} will tell \code{LG_save} where to store the
    results, and the result will in addition also contained an
    updated version of \code{info} a \code{bookmark} and the path
    to the info-file, in order for the parent-function to update
    the info-file _after_ the computation has been successfully
    performed and saved.
}
\description{
This is an internal function that works upon the \code{spy}-report
from the targeted function, that should be one of the six functions
\code{LG_approx_scribe}, \code{LG_spectra_scribe},
\code{LG_boot_approx_scribe}, \code{LG_boot_spectra_scribe},
\code{LG_Wrapper_Original} and \code{LG_Wrapper_Bootstrap}.
}
\details{
This function starts out by calling
    \code{LG_sanity_checks} that investigates if the arguments
    given to the targeted function makes sense.  Then it does as
    its name suggests, i.e. it takes care of the bookkeeping of the
    information stored in the info-file.
}
