% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_boot_approx.R
\name{LG_boot_approx}
\alias{LG_boot_approx}
\title{Approximations of the local Gaussian auto- and cross-correlations for bootstrapped replicates}
\usage{
LG_boot_approx(
  save_dir = NULL,
  TS_boot,
  lag_max,
  LG_points,
  .bws_mixture,
  bw_points,
  .bws_fixed,
  .bws_fixed_only,
  content_details,
  LG_type
)
}
\arguments{
\item{save_dir}{A specification of the directory to be used when
saving (and loading) data.  The default value \code{NULL} will
turn of the interaction with disk.}

\item{TS_boot}{The matrix of bootstrap-replicates produced by
\code{TS_boot_sample}, where we want to find estimated
parameters of the local Gaussian approximations.}

\item{lag_max}{The number of lags to include in the analysis.}

\item{LG_points}{An array that specifies the point at which it is
desired to compute the local Gaussian estimates.  The default
value \code{NULL} will imply that the values used in the
computations upon the original data are recycled.  Values can
alternatively be computed by the help of the function
\code{LG_select_points}, but these must then be a subset of the
original values.}

\item{.bws_mixture}{An argument that specifies how the global
bandwidths and those obtained by the nearest-neighbour strategy
should be combined.  The three available options are
\code{c("mixture", "local", "global")}, which have the
following effects.  The alternatives \code{local} and
\code{global} will respectively only select the nearest
neighbour or global.  These alternatives seems however to not
work well when used on some parts of the lagged pairs of
interest, i.e. the nearest neighbour values might be to "small"
in the center of the distribution, whereas the global
bandwidths seems to fail in the periphery of the distribution.
The alternative \code{mixture} attempt to resolve this by (for
each grid point) selecting the largest of the two alternative
bandwidths.  Note that the value of \code{.bws_mixture} decides
how much information that is computed, i.e. the alternative
\code{local} will turn off the computation of global
bandwidths.  However, the computations of the nearest neighbour
bandwidths will also be computed when the alternative
\code{global} is used, since it does not take long to compute
and it is that function that creates the array we need as a
mould for the result.  If the user does not make a selection,
then all three alternatives will be computed.}

\item{bw_points}{A vector, default \code{c(25, 35)}, that specifies
the percentage of the observations that we want inside the
"bandwidth-square".  If \code{.bws_mixture} is selected to be
\code{global}, then this argument will be ignored. and no
nearest neighbours will be computed.}

\item{.bws_fixed}{A vector of non-negative real values, that can be
used to specify fixed values for the bandwidths (which might be
of interest to do in a preliminary analysis).  The default
value \code{NULL} will prevent the computation of Local
Gaussian Estimates based on fixed bandwidths.}

\item{.bws_fixed_only}{A logic value, default \code{FALSE}, that
can be used to drop the rather time-consuming data-driven
estimation of bandwidths in favour of the simplified approach
where fixed bandwidths are used instead.  Note that
\code{.bws_fixed} must be specified when \code{.bws_fixed_only}
are set to \code{TRUE}.}

\item{content_details}{A character string from \code{c("rho_only",
"rho_log.fun", "rho_all")}, of which the first will be selected
if the user doesn't specify something else.  The purpose of
this argument is to decide the amount of details that should be
stored from the estimates governed by the argument
\code{LG_type}.  The default \code{rho_only} is the one needed
for the computation of the Local Gaussian Spectral Densities,
whereas \code{rho_log.fun} in addition returns the estimated
values of the logarithm of the density.  The option
\code{rho_all} will in addition return all the parameters when
\code{LG_type} is given as \code{par_five}.}

\item{LG_type}{One of \code{c("par_five", "par_one")}, i.e. should the
Local Gaussian Autocorrelations be based on the approach using
five parameters or the simplified approach using only one
parameter.  The default choice is \code{"par_five"}.}
}
\value{
This function is a wrapper around \code{LG_approx}, and the
    result is thus the same as specified for that function.
}
\description{
This function is in essence a wrapper for the function
    \code{LG_approx}, and it is called when we need to investigate
    bootstrap-based replicates in our analysis of the local
    Gaussian spectral densities.
}
\details{
This function can be called manually, but the intention is
    that it only should be called from
    \code{LG_boot_approx_scribe}, since that will ensure that the
    arguments are properly recorded and that the result are saved
    to appropriately named files.  In order to dissuade users from
    calling this (often quite time consuming) function directly, no
    default values have been specified for the arguments.
}
\note{
Regarding the case where the \code{LG_type}-argument is equal
    to "par_one": The author of this package has always considered
    the "par_one"-approach to be reasonable when the aim of the
    investigation is to estimate a density at a given point.
    However, the extraction of the correlation value from the
    resulting density-estimate will in general not capture the
    local geometrical properties of the targeted distribution at
    the point of investigation.  The "par_one"-approach is as such
    (in general) a complete waste of computation resources.
}
