% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_create_plot_df.R
\name{LG_create_plot_df}
\alias{LG_create_plot_df}
\title{Create the data needed when plotting the local Gaussian
correlations and spectra.}
\usage{
LG_create_plot_df(look_up, ..env)
}
\arguments{
\item{look_up}{The list created by \code{LG_lookup}, containing
the information needed in order to decide what kind of
data-frame that is required.}

\item{..env}{The environment containing the required information,
and also the environment where the resulting environment will
be stored.}
}
\value{
This function adds a new environment to the environment
    \code{..env}, which contains the data needed for the function
    \code{LG_plot} to work as desired.
}
\description{
This function created an environment with the data needed for the
plots of the estimated local Gaussian correlations and spectra.  A
simple caching is performed in order to avoid having computations
more than once.
}
\keyword{internal}
