% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGSD.R
\docType{data}
\name{LG_default}
\alias{LG_default}
\title{Internal default values.}
\format{An object of class \code{list} of length 34.}
\usage{
LG_default
}
\description{
This list contains internal default-values, that other functions in
the package need in order to "communicate" together in a simple
manner.  That is to say, this list governs the naming conventions
used upon dimension-names for arrays; it specifies the
directory-hierarchy and the prefixes to be used for the the files
stored in them; and it also specifies the information that the
sanity-check function need to have access to.
}
\details{
THIS DOCUMENTATION HAS TO BE UPDATED!  See the use of @format and
"describe" for the documentation of \code{TS_families}.

The present stuff contained in \code{LG_default} are:

\code{boot.prefix}, that are used to name the arrays that contain
bootstrap-related information, and in the functions that later on
extract information from them.

\code{boot.dir}, that will be used as prefix for the
bootstrap-directories, such that different settings for the
bootstrap-procedure can thus be compared against each other.

\code{approx.save}, three file-names to be used when storing
estimated parameters/function values from the local Gaussian
approximations for the bootstrap-replicates. Two files are used for
the data, due to the difference in structure between the zero-lag
case and the rest.  The third file is used to store information to
be used for the extraction of the data needed for the construction
of the local Gaussian spectra, i.e. the paths to the two first
files, and quotes that can be evaluated in order to extract the
information needed for the computation of the local Gaussian
spectra. In addition to this there is a numeric summarising the
e-flag status for the values computed by \code{localgauss}, this
should be zero if no problems occurred.

\code{boot.approx.save}, that gives the three file-names to be used
when storing estimated parameters/function values from the local
Gaussian approximations for the bootstrap-replicates. Two files are
used for the data, due to the difference in structure between the
zero-lag case and the rest.  The third file is used to store
information to be used for the extraction of the data needed for
the construction of the local Gaussian spectra, i.e. the paths to
the two first files, specification of the subsetting to be used and
a vector with the e-flag status for the values computed by
\code{localgauss}.

\code{boot.spectrals.save}, that gives the file-name to be used when
storing the derived data based on the data from
\code{LG_boot_approx}.
}
\keyword{internal}
