% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup.R
\name{LG_lookup}
\alias{LG_lookup}
\title{Look up stuff from the information object.}
\usage{
LG_lookup(.info, .selected, .path_selection_list, .input = NULL)
}
\arguments{
\item{.info}{The information object central for all the bookkeeping.}

\item{.selected}{The key that identifies the part that should be
visualised.}

\item{.path_selection_list}{A list that together with
\code{.selected} enables the extraction of the relevant
information from \code{.info}}

\item{.input}{The (non-reactive version) of the \code{.input}-list
generated by \code{shiny} can be added here, and when that is
done a revised version of the list will be returned to the
workflow.  This revised version will be the basis for the
plot-attribute to be used when writing dynamic documents.  The
default value \code{NULL} will ignore this update.}
}
\value{
This function will update the environment of the calling
    function.  If \code{.input} is different from \code{NULL}, then
    a revised version of that will be returned to the workflow
    (under the name \code{.input}).  (The revised version will not
    change any of the values in \code{.input}, it will rather add
    new sublists to be used by other functions later on.)  In the
    other cases, the following objects will be added to the
    environment of the calling function: Three logical values
    \code{.block}, \code{.only_diagonal} and \code{.multivariate},
    three values related to points under investigation
    \code{.Horizontal}, \code{.Vertical} and \code{.Shape}, and
    finally \code{info_part} that will be used to create the
    interface.
}
\description{
This function looks up details that is needed in several functions
related to the interactive investigation, and was thus created to
avoid the duplication of code.
}

