% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup.R
\name{LG_lookup}
\alias{LG_lookup}
\title{Lookup local Gaussian details based on the logged information}
\usage{
LG_lookup(input, .AB_env)
}
\arguments{
\item{input}{A non-reactive version of the \code{input}-list
generated by \code{LG_shiny}.}

\item{.AB_env}{The environment that contains the desired
information.}
}
\value{
An environment will be returned to the workflow, in which
    there will be a basic component as specified under details, and
    then there are four lists with specialised information as
    outlined below:

\describe{

\item{restrict}{This part is created by the help of
    \code{LG_lookup_restrict}, and it contains the information that
    is needed in order to extract the desired local Gaussian auto-
    and cross-correlations from the data-structure created by the
    scribe-functions.}

\item{cache}{This part is created by the help of
    \code{LG_lookup_cache}, and it contains the unique cache-keys
    that enables a primitive caching-procedure to be used in the
    interactive \code{LG_shiny}-application.}

\item{details}{This part is created by the help of
    \code{LG_lookup_details}, and it contains the information that
    is needed in order to describe the content of a given plot.
    This information will be added as an attribute when a plot is
    created in a nonreactive setting, which implies that it is
    easily available when a plot is to be included in a
    paper/presentation.  See \code{LG_explain_plots} for further
    details.}

\item{curlicues}{This part is created by the help of
    \code{LG_lookup_curlicues}, and it contains the details needed
    in order to annotate assorted information (curlicues) to the
    final plots.  A plot-stamp that reveals the content of the plot
    will always be present together with information about the
    length of the sample and the number of replicates.  Additional
    information will be added depending on the type of plot, this
    can e.g. be details related to the truncation level, the point
    of investigation, the bandwidth, the status for the numerical
    convergence of the estimates, or the block length used for the
    bootstrapping algorithm.  See the scripts for examples related
    to how the annotated text can adjusted when the plots are
    created in a nonreactive setting.}

}
}
\description{
This internal function creates a lookup-environment
    that combines different values from the logged information.
    The purpose is to collect the results of different minor
    computations, which is needed in several functions.
}
\details{
This function will copy the values from \code{input}, and
    supplement those with values extracted from the information
    logged in \code{.AB_env}.  Assorted logical values are computed
    from this, and some minor chunks of derived information is also
    created in the initial stage.  These will be used later on in
    the functions involved in the interactive investigation of the
    different plots that can be created from the estimated local
    Gaussian correlations.  Some additional specialised information
    is then added by the help of related \code{lookup}-functions.
}
\keyword{internal}
