% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_cache.R
\name{LG_lookup_cache}
\alias{LG_lookup_cache}
\title{Helper-function to deal with the caching.}
\usage{
LG_lookup_cache(look_up)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}
}
\value{
A list with cache-values will be returned to the calling
    function \code{LG_lookup}, and it will there be added as a node
    to the \code{look_up}-list.  The cache-values in this list will
    be used in order to avoid that already performed tasks are
    recomputed.
}
\description{
Helper-function to deal with the caching.
}
\keyword{internal}
