% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_lookup_details.R
\name{LG_lookup_details}
\alias{LG_lookup_details}
\title{Helper-function to deal with the details.}
\usage{
LG_lookup_details(look_up, .AB_env)
}
\arguments{
\item{look_up}{A list from the internal workings of the function
\code{LG_lookup}.}

\item{.AB_env}{The environment that contains the desired
information.  This argument is inherited from \code{LG_lookup}.}
}
\value{
A \code{details} list will be returned to the calling
    function (i.e. \code{LG_lookup}), and this list contains
    details needed for the description of the content of the plots.
}
\description{
Helper-function to deal with the details.
}
\keyword{internal}
