% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_normalisation_adjustment.R
\name{LG_normalisation_adjustment}
\alias{LG_normalisation_adjustment}
\title{Create a normalised time series from the original ones.}
\usage{
LG_normalisation_adjustment(TS, .adjustment_rule = 0, .remove_ties = TRUE)
}
\arguments{
\item{TS}{The time series object to normalise.  It's assumed that
this is structured in an array, with one dimension-name being
"content".}

\item{.adjustment_rule}{A non-negative number that will be stored
as an attribute that later on can be used in order to add some
additional scaling (in addition to the one from the windows
function) when the estimated local Gaussian auto- and
cross-correlations are used for the computation of the local
Gaussian spectra.  The default value \code{0} ensures that only
the scaling from the window-function is used.}

\item{.remove_ties}{A logical value, default \code{TRUE}, in which
case the presence of ties will trigger a minor perturbation of
the data.  Note: Whenever this happens, \code{set.seed(1)}
will be used in order to ensure reproducibility.}
}
\value{
A normalised version of the time series of interest.
}
\description{
Create a normalised time series from the original ones.
}
\details{
A normalisation of the time series under consideration is
    preferable when an analysis based on the local Gaussian
    approximations are of interest.  The intention of this function
    is to allow for easier adjustment of the normalisation regime,
    i.e if the normalisation should be based on the cumulative
    density function (or other options that might be added later
    on).  This function will also add some attributes to the
    result, with the required instruction for the finite-sample
    adjustments to be used when computing the local Gaussian
    autocorrelations.
}

