% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot.R
\name{LG_plot}
\alias{LG_plot}
\title{Create plots to inspect different aspects of Local Gaussian
Spectral Densities.}
\usage{
LG_plot(.data, .global = NULL, .lag = NULL, .percentile = NULL,
  .select = c("all", "canvas", "add"), .canvas = NULL, .title = NULL,
  .xlim = c(0, 0.5), .ylim = NULL, .aes_xy = NULL, .aes_min_max = NULL,
  .sanity_checks = TRUE)
}
\arguments{
\item{.data}{A data-frame containing the desired information needed
by the argument \code{data} in the function \code{ggplot}.}

\item{.global}{A list with the two components \code{.block} and
\code{.data}, that can be used to add information about the
ordinary spectral density.  The default value \code{NULL} turns
of this part of the plot.}

\item{.lag}{A character-string, which when specified will add the
lag that is under inspection.  The default value \code{NULL}
turns of this part of the plot.}

\item{.percentile}{A character-string, which when specified will
add the percentile under investigation (in the upper right
corner).  The default value \code{NULL} turns of this part of
the plot.}

\item{.select}{An argument from the vector \code{c("all", "canvas",
"add")} that selects the desired plot.  The first alternative
is the one selected by default if no specification is made by
the user.  The two other alternatives are intended used when
this function is called from \code{\link{LG_shiny}}, in order
to reduce the amount of data that must be updated in order to
create the interactive visualisation.}

\item{.canvas}{To be used when adding details to a plot produced
earlier on.  This argument is needed when \code{LG_plot} is
called from \code{\link{LG_shiny}}.  The default value
\code{NULL} can be left unchanged in the general case.  Note
that this argument will be ignored unless \code{.select} is
given as \code{add}.}

\item{.title}{The title to be used in the plot.}

\item{.xlim}{Numeric vector of length two, giving the
\code{xlim}-value to be used in \code{ggplot}.  The default
value \code{c(0, 0.5)} has been selected for the visualisation
of the local Gaussina spectra.  The value \code{NULL} can be
used to allow \code{ggplot} to select reasonable values based
on the values in \code{.data}.}

\item{.ylim}{Numeric vector of length two, giving the
\code{ylim}-value to be used in \code{ggplot}.  The default
value \code{NULL} will allow \code{ggplot} to select reasonable
values based on the values in \code{.data}.}

\item{.aes_xy}{A call of \code{aes} that will be used \code{ggplot}
requires an \code{aes} specifying \code{x} and \code{y}.  The
default value \code{NULL} will trigger an error if unspecified.}

\item{.aes_min_max}{A quoted expression to be used when calling
\code{ggplot} with a mapping that requires us to give
\code{aes} with a specification of \code{min} and \code{max}.
The default value \code{NULL} will stop the execution of any
layers depending on \code{min} and \code{max}.}

\item{.sanity_checks}{Logic value, default \code{TRUE}, that
decides whether or not a sequence of sanity-checks will be run
on the arguments.  When used in \code{LG_shiny}, it's
preferable to be able to turn of these tests.}
}
\value{
The first incarnation of this code will focus on the
    creation of plots depicting the local Gaussian spectral
    densities, with our without the information required to create
    the bootstrap based confidence intervals.  Other plots will
    probably be added later on.
}
\description{
To get a joint setup for plots used in interactive sessions with
\code{LG_shiny} and plots used in a more static setting, the
intention of this function is to collect the different plots in one
main function.
}
\details{
In most cases, the data will be extracted from one or more files
into an array (due to memory restraints we might need a solution
that can do this in several parts) and based on a collection of
inputs we will then need to extract a relevant part into a
data-frame in order for ggplot to have the desired format to work
upon.  In this first iteration of the code, I think I will restrict
attention to the case where we assume the data already has been
extracted.
}

