% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot.R
\name{LG_plot}
\alias{LG_plot}
\title{Create plots to inspect different aspects of Local Gaussian
Spectral Densities and Local Gaussian Correlations.}
\usage{
LG_plot(.data_list, .lag = NULL, .percentile = NULL,
  .select = c("all", "canvas", "add"), .canvas = NULL, .title = NULL,
  .xlim = c(0, 0.5), .ylim = NULL, .aes_list = NULL)
}
\arguments{
\item{.data_list}{A list that with the data-frames that contains
the desired information needed by the \code{data}-argument in
the function \code{ggplot}.}

\item{.lag}{A character-string, which when specified will add the
lag that is under inspection.  The default value \code{NULL}
turns of this part of the plot.}

\item{.percentile}{A character-string, which when specified will
add the percentile under investigation (in the upper right
corner).  The default value \code{NULL} turns of this part of
the plot.}

\item{.select}{An argument from the vector \code{c("all", "canvas",
"add")} that selects the desired plot.  The first alternative
is the one selected by default if no specification is made by
the user.  The two other alternatives are intended used when
this function is called from \code{\link{LG_shiny}}, in order
to reduce the amount of data that must be updated in order to
create the interactive visualisation.}

\item{.canvas}{To be used when adding details to a plot produced
earlier on.  This argument is needed when \code{LG_plot} is
called from \code{\link{LG_shiny}}.  The default value
\code{NULL} can be left unchanged in the general case.  Note
that this argument will be ignored unless \code{.select} is
given as \code{add}.}

\item{.title}{The title to be used in the plot.}

\item{.xlim}{Numeric vector of length two, giving the
\code{xlim}-value to be used in \code{ggplot}.  The default
value \code{c(0, 0.5)} has been selected for the visualisation
of the local Gaussian spectra.  The value \code{NULL} can be
used to allow \code{ggplot} to select reasonable values based
on the values in \code{.data}.}

\item{.ylim}{Numeric vector of length two, giving the
\code{ylim}-value to be used in \code{ggplot}.  The default
value \code{NULL} will allow \code{ggplot} to select reasonable
values based on the values in \code{.data}.}

\item{.aes_list}{A list containing the different \code{aes}-expressions
needed for the plots.}
}
\value{
The first incarnation of this code will focus on the
    creation of plots depicting the local Gaussian spectral
    densities, with our without the information required to create
    the bootstrap based confidence intervals.  Other plots will
    probably be added later on.
}
\description{
To get a joint setup for plots used in interactive sessions with
\code{LG_shiny} and plots used in a more static setting, the
intention of this function is to collect the creation of the
different plots in one main function.  Note that the arguments of
this function is not as such intended to be written out by the user
when the function is used in a static setting, instead they are
supposed to be given from the \code{LG_shiny}-interface.
}
