% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_complex.R
\name{LG_plot_complex}
\alias{LG_plot_complex}
\title{Complex-valued-plots for LGSD-data}
\usage{
LG_plot_complex(..env, look_up, .selection = c("polar", "Cartesian",
  "zoom"))
}
\arguments{
\item{..env}{The environment containing the desired information
from which the data should be extracted}

\item{look_up}{The usual list of details needed for the
investigation.}

\item{.selection}{Specifies the type of plot to be used.  This
should be a temporary argument, as it seems natural that it
should be defined as a part of the shiny-interface.}
}
\value{
A heatmap-plot of either the estimated local Gaussian
    spectral density estimates for a combination of
    points-frequency, or points-lags.
}
\description{
Complex-valued-plots for LGSD-data
}
\keyword{internal}
