% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_complex.R
\name{LG_plot_complex}
\alias{LG_plot_complex}
\title{Complex-valued-plots for the local Gaussian spectra}
\usage{
LG_plot_complex(..env, look_up)
}
\arguments{
\item{..env}{The environment containing the desired information
from which the data should be extracted}

\item{look_up}{The environment containing the details needed for
the investigation.}
}
\value{
A plot is returned to the workflow.  This plot shows the
    complex-valued estimates of the local Gaussian spectra for the
    given combination of frequency, point, truncation level and
    bandwidth.  The resulting (loud of )points can be presented in
    a zoomed in version, or they can be inspected in a zoomed out
    version with confidence intervals that can be based either on a
    Cartesian or polar representation of the complex numbers.
}
\description{
This internal function plots the complex-valued
    spectra (for a given frequency) in the complex plane.
}
\keyword{internal}
