% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_heatmap.R
\name{LG_plot_heatmap}
\alias{LG_plot_heatmap}
\title{Heatmap-plots for LGSD-data}
\usage{
LG_plot_heatmap(..env, look_up)
}
\arguments{
\item{..env}{The environment containing the desired information
from which the data should be extracted}

\item{look_up}{The usual list of details needed for the
investigation.}
}
\value{
A heatmap-plot of either the estimated local Gaussian
    spectral density estimates for a combination of
    points-frequency, or points-lags.
}
\description{
Heatmap-plots for LGSD-data
}
\keyword{internal}
