% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_helper.R
\name{LG_plot_helper}
\alias{LG_plot_helper}
\title{Helper for \code{LG_plot}.}
\usage{
LG_plot_helper(main_dir, input, show_graphical, path_selection_list, .env)
}
\arguments{
\item{main_dir}{The part of the \code{data_dir}-argument from the
calling function \code{LG_shiny}, that specifies the main path
to where the information is stored.}

\item{input}{A list containing the arguments needed in order to
perform the extraction.}

\item{show_graphical}{A logical vector that specifies the kind of
data that should be investigated.  Required in order to
identify the correct array to work upon (since more than one
can be available).}

\item{path_selection_list}{A list containing the components needed
in order to identify the correct array to work upon (since more
than one can be available).}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}
}
\value{
A plot will be returned.
}
\description{
This is a helper function for \code{LG_plot}, that first extracts
the desired array from file and then creates the plot of interest.
The function are used as an helper in \code{LG_shiny}, in which
case it also will take care of the restrictions in a manner that
attempts to avoid the major restrictions when possible.  The
interactive setting can trigger the creation of code needed in
order to recreate a plot in a non-reactive setting too (e.g. as an
example in a paper/presentation), in which case some of the
internal workings of the function will be adjusted to allow for the
function call to work without the \code{shiny}-package being
loaded.
}

