% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_plot_helper.R
\name{LG_plot_helper}
\alias{LG_plot_helper}
\title{Helper-function for \code{LG_plot}.}
\usage{
LG_plot_helper(main_dir, input, .env, .extract_LG_data_only = FALSE)
}
\arguments{
\item{main_dir}{The part of the \code{data_dir}-argument from the
calling function \code{LG_shiny}, that specifies the main path
to where the information is stored.  Note that this argument
only is required when the function is called outside of the
shiny application.  (Inside of the shiny-application, the
relevant information will be extracted from \code{.env}.)}

\item{input}{A list with the arguments needed in order to extract
the desired data for the plot of interest.}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}

\item{.extract_LG_data_only}{Logical argument, default
\code{FALSE}.  The default value ensures that a plot is
created, whereas the value \code{TRUE} is used when the desired
outcome is an object containing the estimated local Gaussian
values.}
}
\value{
The default behaviour is that a plot-object will be
    returned, with attributes describing the content. The
    information can be extracted from the plot-object by the help
    of \code{LG_explain_plot}.  When \code{.extract_LG_data_only}
    is \code{TRUE}, the result will be an object containing the
    estimated local Gaussian values.
}
\description{
This is a helper function for \code{LG_plot}, that first extracts
the desired array from file and then creates the plot of interest.
The function are used as an helper in \code{LG_shiny}, in which
case it also will take care of the restrictions in a manner that
attempts to avoid the major restrictions when possible.  The
interactive setting can trigger the creation of code needed in
order to re-create a plot in a non-reactive setting too (e.g. as an
example in a paper/presentation), in which case some of the
internal workings of the function will be adjusted to allow for the
function call to work without the \code{shiny}-package being
loaded.
}
\details{
NOTE: This function is exported in order for interesting plots to
be recreated outside of the shiny-application, in which case the
returned plot-objects also will contain attributes needed for the
proper identification of the content of the plots.  The idea is
that the required parameters of this function will be filled in by
the shiny-application, and even though it is possible to adjust the
input parameters manually, it is not recommended.
}
