% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_sanity_cheks.R
\name{LG_sanity_checks}
\alias{LG_sanity_checks}
\title{Check the sanity of arguments to the scribes and the wrappers}
\usage{
LG_sanity_checks(spy_report)
}
\arguments{
\item{spy_report}{The result returned by \code{spy} when used on
the targeted function.}
}
\value{
If any arguments are erroneous, the program will be stopped
    and information about the problems returned to the workspace.
    If no problems are found, the programs returns nothing.  Note
    that the program will work upon a shortcut to the environment
    part of \code{spy_report}, and that implies that any updates
    will have been registered in the \code{spy_report} in the
    calling function.
}
\description{
This function checks that the arguments are valid, and it might
also add values to the environment of the function under
investigation.
}
\keyword{internal}
