% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_explain_interface.R
\name{LG_shiny_explain_interface}
\alias{LG_shiny_explain_interface}
\title{Create the instructions for \code{LG_shiny}.}
\usage{
LG_shiny_explain_interface(input, .env, .env2)
}
\arguments{
\item{input}{A list containing the arguments needed in order to
perform the extraction.}

\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
An explanation will be returned.
}
\description{
This internal function is a helper-function for
    \code{LG_shiny}, that creates some explanations that can be
    added to the interface.  This version is still in its infancy,
    i.e. it is mostly a placeholder.
}
\keyword{internal}
