% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_explain_plots.R
\name{LG_shiny_explain_plots}
\alias{LG_shiny_explain_plots}
\title{Explain details about the plots created by \code{LG_shiny}.}
\usage{
LG_shiny_explain_plots(input, .env)
}
\arguments{
\item{input}{A list containing the arguments needed in order to
perform the extraction.}

\item{.env}{An environment in which \code{.arr} will be updated.
This argument can be skipped when called in a non-interactive
setting.}
}
\value{
An explanation will be returned.
}
\description{
This helper function for \code{LG_shiny} creates text that can
explain the content of the plots, i.e. details related to whether
the data is simulated or based on observations, and some summaries
with regard to the parameters used in the computation.
}
\keyword{internal}
