% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LG_shiny_S_input.R
\name{LG_shiny_S_input}
\alias{LG_shiny_S_input}
\title{Create the spectrum-specific part of the \code{LG_shiny}-interface.}
\usage{
LG_shiny_S_input(.env, .env2)
}
\arguments{
\item{.env}{The environment where the original arguments given to
\code{LG_shiny} lives, i.e. arguments like \code{main_dir} and
\code{data_dir}.}

\item{.env2}{The environment containing the two lists \code{input}
and \code{output}.}
}
\value{
This function creates the \code{actionButtons} needed for
    the spectrum-specific part of the \code{LG_shiny}-interface.
    This includes the selection of the type of spectrum to
    investigate (auto- or cross-spectrum, with possible
    complex-valued results), and the arguments to be used when
    computing the different spectra (frequency-range, truncation
    point and lag-window-function).
}
\description{
This internal function is used to update the interface
    created by \code{LG_shiny}, in order for the investigation to
    focus on the local Gaussian (auto- and cross-) spectra.
}
\note{
This function creates the two output-nodes, "spectrum_type"
    and "spectrum_arguments".  Both nodes will be created in the
    initial phase, whereas only the "spectrum_type" needs to be
    updated later on (since different spectra will occur depending
    on auto- and cross-versions of the spectra, and points on or
    off the diagonal).  The internal workings of this function thus
    requires a comparison of old and new input-parameters.  Note
    that all of these are equal in the initiation phase.
}
\keyword{internal}
