% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_boot_sample.R
\name{TS_boot_sample}
\alias{TS_boot_sample}
\title{Create bootstrap-replicates of a time series.}
\usage{
TS_boot_sample(TS, main_dir, save_dir, nb = 5, boot_type = "block",
  block_length = 20, boot_seed = NULL)
}
\arguments{
\item{TS}{The original time series that we want to create
bootstrap-replicates of.}

\item{main_dir}{The path to the main directory, that contains the
file-hierarchy created when using the local Gassian approach
for the investigation of time series.}

\item{save_dir}{A specification of the directory to be used when
saving (and loading) data.  The default value \code{NULL} will
turn of the interaction with disk.}

\item{nb}{An integer that specifies how many bootstrap-replicates
we want to use in our analysis.  Default value 5 (at least in
the development phase).}

\item{boot_type}{This one can be used to select what kind of
bootstrap-algorithm to be used.  Default value \code{"block"},
but it has not been implemented any other algorithms
yet... (Reminder: It would be be nice to include the
'network-duality' approach in this context -- but then I fear I
might need to implement that routine from scratch.)}

\item{block_length}{The length of the blocks to be used when
\code{boot_type="block"} is used.  Default value 20 in the
development phase, but I suppose in general it should be some
formula based on the time series under investigation.}

\item{boot_seed}{Use this to enable reproducible results.  The
default value \code{NULL} will trigger a random seed to be
selected for this value (that then will be recorded in case a
reproduction of the result is desired later on).}
}
\value{
The result of this function will be a matrix with \code{nb}
    rows and \code{length(TS)} columns, i.e.  each row is a
    bootstrapped replicate of \code{TS}.
}
\description{
Based on some algorithm for bootstrapping of a time series, a
matrix of bootstrapped series are returned.  The returned object
does also include an attribute specifying the adjustment-terms
needed to shift from estimates of covariances and correlations for
the lagged pairs to the estimates of autocovariances and
autocorrelations.
}

