% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_cibbb_tuples.R
\name{TS_cibbb_tuples}
\alias{TS_cibbb_tuples}
\title{Circular index-based block bootstrap for tuples}
\usage{
TS_cibbb_tuples(.indices, .lag)
}
\arguments{
\item{.indices}{The initial vector of (resampled) indices.}

\item{.lag}{The lag-value that should be used.}
}
\value{
This helper function returns a list with two vectors, one
    named \code{first} and one named \code{second}.  These are the
    indices needed for the circular index-based block bootstrap for
    tuples.
}
\description{
This internal function creates the bivariate pairs of
    indices needed in order for the circular index-based block
    bootstrap for tuples to be used in the computations
}
\keyword{internal}
