% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_load.R
\name{TS_load}
\alias{TS_load}
\title{Load saved file and return \code{TS} and \code{save_dir}}
\usage{
TS_load(.TS_info, save_dir = TRUE)
}
\arguments{
\item{.TS_info}{A list containing the three components \code{TS},
\code{main_dir} and \code{save_dir}.}

\item{save_dir}{Logical value, default \code{TRUE}.  To be used
when it's of interest to have a path at which the next step in
the computation should be stored.}
}
\value{
This function will add objects to the calling environment.
}
\description{
This is an internal helper-function that takes care of some pesky
details when loading and saving arrays related to time series.  The
point is that regardless of the name used upon the stored array, it
will be returned to the calling environment with the name
\code{TS}, and it's also possible to get information required for
the saving of the results to be computed from \code{TS}.
}

