% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_sample.R
\name{TS_sample}
\alias{TS_sample}
\title{Time Series Sample-Generator}
\usage{
TS_sample(TS_key = "WNG", N = 500, nr_samples = 1, ...,
  .seed = NULL, .kind_vstr_list = NULL)
}
\arguments{
\item{TS_key}{A key, i.e. a \code{character}, corresponding to an
element in \code{TS_families}.  Default value \code{WNG},
i.e. "White Noise Gaussian".}

\item{N}{The desired length of the time series.  Default value
\code{500}.}

\item{nr_samples}{The desired number of independent samples to be
produced.  The default value \code{1} is the one to use when
testing for the effect of different settings for bootstrapping,
whereas (much) higher values is intented to be used to get an
idea about what the true behaviour should be when investigating
the Local Gaussian Spectral Density.}

\item{...}{\code{dotsMethods}-strategy for feeding parameters to
the function (identified by \code{TS_key} that generates the
time series.}

\item{.seed}{Use this to enable reproducible results.  Default value
\code{NULL} (it will be generated and recorded in the code).}

\item{.kind_vstr_list}{This can be used to create a list with the
values for \code{kind}, \code{normal.kind} and \code{vstr}.
(See the help-page of \code{Random} for details about these
three arguments.)  Note that the default value \code{NULL} will
imply that the function \code{set_seed} will be used to create
the required list based on the present settings.}
}
\value{
This function returns a list with four parts \code{TS},
    \code{TS_data}, \code{spy_report} and \code{seed_vec}.
    \code{TS} contains an array with the time series generated
    according to the specified arguments, whereas \code{TS_data}
    includes some additional stuff that will be used when
    \code{TS_LG_object} takes care of the saving of the data. The
    arguments needed in order to redo the computation later on is
    stored in \code{spy_report}.  The part \code{seed_vec} reflects
    that the internal workings of the code creates a vector of
    seeds (based on \code{.seed} that then can be used to create an
    individual series later on (this is of course only of interest
    when \code{nr_samples} is larger than one, and we for some
    reason later on would like to do an in depth analysis of one of
    the resulting time series.
}
\description{
Time Series Sample-Generator
}
\details{
This function will create one or more time series based on
    the "keys" stored in \code{TS_families}, with emphasis on also
    storing the required arguments needed in order to re-create it
    later on.
}
